### 2022-06-10 505e4d3
Update ddpm.py
Co-Authored-By: Hasan Muhammad Sayeed <47003757+hasan-sayeed@users.noreply.github.com>
