### 2022-06-10 ea3e7c2
use augment_horizontal_flip=False and save to unique results folder
