### 2022-06-11 5b4008e
Create ddpm_pretrained_sample.py
