### 2022-06-11 6da1cd9
add assert test for space groups matching
