### 2022-06-11 feb2161
Revert "add assert test for space groups matching"
This reverts commit 6da1cd915cb79496d2f4ff4ebaf579334964a6c7.
