### 2022-06-11 e741427
use SpacegroupAnalyzer to detect symmetry of decoded structures
