### 2022-06-11 aa1c256
initial, loose check with StructureMatcher followed by detailed match
use `get_s2_like_s1` for the detailed matching

https://pymatgen.org/pymatgen.analysis.structure_matcher.html#pymatgen.analysis.structure_matcher.StructureMatcher.get_s2_like_s1
