### 2022-06-11 acd9c02
refined_example_structures
