### 2022-06-11 491ad48
Revert "refined_example_structures"
This reverts commit acd9c026d2be5f8961ff58360e532bf3c0f741e7.
