### 2022-06-13 9f26904
basic port of notebook functionality (WIP)
Co-Authored-By: Hasan Muhammad Sayeed <47003757+hasan-sayeed@users.noreply.github.com>
