### 2022-06-13 c5dd807
add attributions section
@michaeldalverson
