### 2022-06-14 323fc5d
fix mistake where low_val should have been upp_val
