### 2022-06-14 fc8e72c
fit method and assert tests
atomic_numbers use np.min/max instead of quantiles
space_group use np.min/max instead of quantiles
rest use quantiles
assert tests is just fixing it to its current output. Need to go back and manually verify this.
