### 2022-06-14 f88c0ff
fixup: loop through stacked distance matrices
