### 2022-06-14 2208564
change to for loop to avoid mypy warning
