### 2022-06-14 dbbcaeb
encode vs. decode symprec, angle_tolerance, and primitive options
