+++ src/xtal2png/core.py
@@ -105,14 +105,24 @@ class XtalConverter:
     save_dir : Union[str, 'PathLike[str]']
         Directory to save PNG files via ``func:xtal2png``,
         by default path.join("data", "interim")
-    symprec : float, optional
+    symprec : Union[float, Tuple[float, float]], optional
         The symmetry precision to use when decoding `pymatgen` structures via
-        ``func:pymatgen.symmetry.analyzer.SpaceGroupAnalyzer.get_refined_structure``. By
-        default 0.1.
-    angle_tolerance : Union[float, int], optional
+        ``func:pymatgen.symmetry.analyzer.SpaceGroupAnalyzer.get_refined_structure``. If
+        specified as a tuple, then ``symprec[0]`` applies to encoding and ``symprec[1]``
+        applies to decoding. By default 0.1.
+    angle_tolerance : Union[float, int, Tuple[float, float], Tuple[int, int]], optional
         The angle tolerance (degrees) to use when decoding `pymatgen` structures via
-        ``func:pymatgen.symmetry.analyzer.SpaceGroupAnalyzer.get_refined_structure``. By
-        default 5.0.
+        ``func:pymatgen.symmetry.analyzer.SpaceGroupAnalyzer.get_refined_structure``. If
+        specified as a tuple, then ``angle_tolerance[0]`` applies to encoding and
+        ``angle_tolerance[1]`` applies to decoding. By default 5.0.
+    encode_as_primitive : bool, optional
+        Encode structures as symmetrized, primitive structures. Uses ``symprec`` if
+        ``symprec`` is of type float, else uses ``symprec[0]`` if ``symprec`` is of type
+        tuple. Same applies for ``angle_tolerance``. By default True
+    decode_as_primitive : bool, optional
+        Decode structures as symmetrized, primitive structures. Uses ``symprec`` if
+        ``symprec`` is of type float, else uses ``symprec[1]`` if ``symprec`` is of type
+        tuple. Same applies for ``angle_tolerance``. By default True
 
     Examples
     --------
@@ -134,8 +144,10 @@ class XtalConverter:
         distance_range: Tuple[float, float] = (0.0, 18.0),
         max_sites: int = 52,
         save_dir: Union[str, "PathLike[str]"] = path.join("data", "preprocessed"),
-        symprec: float = 0.1,
-        angle_tolerance: float = 5.0,
+        symprec: Union[float, Tuple[float, float]] = 0.1,
+        angle_tolerance: Union[float, int, Tuple[float, float], Tuple[int, int]] = 5.0,
+        encode_as_primitive: bool = True,
+        decode_as_primitive: bool = True,
     ):
         """Instantiate an XtalConverter object with desired ranges and ``max_sites``."""
         self.atom_range = atom_range
@@ -149,8 +161,23 @@ class XtalConverter:
         self.distance_range = distance_range
         self.max_sites = max_sites
         self.save_dir = save_dir
-        self.symprec = symprec
-        self.angle_tolerance = angle_tolerance
+
+        if isinstance(symprec, (float, int)):
+            self.encode_symprec = symprec
+            self.decode_symprec = symprec
+        elif isinstance(symprec, tuple):
+            self.encode_symprec = symprec[0]
+            self.decode_symprec = symprec[1]
+
+        if isinstance(angle_tolerance, (float, int)):
+            self.encode_angle_tolerance = angle_tolerance
+            self.decode_angle_tolerance = angle_tolerance
+        elif isinstance(angle_tolerance, tuple):
+            self.encode_angle_tolerance = angle_tolerance[0]
+            self.decode_angle_tolerance = angle_tolerance[1]
+
[…]
