### 2022-06-14 #4465 @rkurchin
👋 @dandavies99 and @trallard, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
