### 2022-06-14 f3f8d31
modify `structures` directly instead of appending to a new list
https://github.com/sparks-baird/xtal2png/issues/98
