### 2022-06-14 5a9c050
same fix for xtal2png top-level
