### 2022-06-14 817487b
ignore type change for `structures`
