### 2022-06-16 7a9d28d
change `structures_to_arrays` syntax and add docstring to process_filepaths_or_structures
