### 2022-06-16 736ed3c
test the refactor of arrays_to-structures
