### 2022-06-16 4ea2189
Update xtal2png_test.py
