### 2022-06-16 69bffa2
get rid of multiple return type finagling and return helpful error message in case of confusion
also delete some commented code
