### 2022-06-16 00d66a2
apply the tol_multiplier to the StructureMatcher too
