### 2022-06-16 7386edd
relax_on_decode kwarg via m3gnet
