+++ src/xtal2png/core.py
@@ -13,6 +13,7 @@ from warnings import warn
 
 import numpy as np
 import pandas as pd
+from m3gnet.models import Relaxer
 from numpy.typing import NDArray
 from PIL import Image
 from pymatgen.core.lattice import Lattice
@@ -125,6 +126,8 @@ class XtalConverter:
         Decode structures as symmetrized, primitive structures. Uses ``symprec`` if
         ``symprec`` is of type float, else uses ``symprec[1]`` if ``symprec`` is of type
         tuple. Same applies for ``angle_tolerance``. By default True
+    relax_on_decode: bool, optional
+        Use m3gnet to relax the decoded crystal structures.
 
     Examples
     --------
@@ -150,6 +153,7 @@ class XtalConverter:
         angle_tolerance: Union[float, int, Tuple[float, float], Tuple[int, int]] = 5.0,
         encode_as_primitive: bool = False,
         decode_as_primitive: bool = False,
+        relax_on_decode: bool = True,
     ):
         """Instantiate an XtalConverter object with desired ranges and ``max_sites``."""
         self.atom_range = atom_range
@@ -180,6 +184,7 @@ class XtalConverter:
 
         self.encode_as_primitive = encode_as_primitive
         self.decode_as_primitive = decode_as_primitive
+        self.relax_on_decode = relax_on_decode
 
         Path(save_dir).mkdir(exist_ok=True, parents=True)
 
@@ -229,7 +234,7 @@ class XtalConverter:
         >>> xc = XtalConverter()
         >>> xc.xtal2png(structures, show=False, save=True)
         """
-        save_names, structures = self.process_filepaths_or_structures(structures)
+        save_names, structures = self.process_filepaths_or_structures(structures)  # type: ignore # noqa: E501
 
         # convert structures to 3D NumPy Matrices
         self.data, self.id_data, self.id_mapper = self.structures_to_arrays(structures)
@@ -270,7 +275,7 @@ class XtalConverter:
         fit_quantiles=(0.00, 0.99),
         verbose=True,
     ):
-        _, structures = self.process_filepaths_or_structures(structures)
+        _, structures = self.process_filepaths_or_structures(structures)  # type: ignore
 
         # TODO: deal with arbitrary site_properties
         atomic_numbers = []
@@ -346,7 +351,7 @@ class XtalConverter:
             setattr(self, name + "_range", tuple(bounds))
 
     def process_filepaths_or_structures(
-        self, structures: Union[List[PathLike], List[Structure]]
+        self, structures: Union[List[str], List[PathLike], List[Structure]]
     ) -> Tuple[List[str], List[Structure]]:
         """Extract (or create) save names and convert/passthrough the structures.
 
@@ -368,11 +373,16 @@ class XtalConverter:
         Raises
         ------
         ValueError
-            _description_
+            "structures should be of same datatype, either strs or pymatgen Structures.
+            structures[0] is {type(structures[0])}, but got type {type(s)} for entry
+            {i}"
         ValueError
-            _description_
+            "structures should be of same datatype, either strs or pymatgen Structures.
+            structures[0] is {type(structures[0])}, but got type {type(s)} for entry
+            {i}"
         ValueError
-            _description_
+            "structures should be of type `str`, `os.PathLike` or
+            `pymatgen.core.structure.Structure`, not {type(structures[i])} (entry {i})"
 
         Examples
         --------
@@ -384,7 +394,7 @@ class XtalConverter:
             if isinstance(s, str) or isinstance(s, PathLike):
                 if first_is_structure:
                     raise ValueError(
-                        f"structures should be of same datatype, either strs or pymatgen Structures. structures[0] is {type(structures[0])}, but got type {type(s)} for entry {i}"  # noqa
[…]
