### 2022-06-17 b1dd62a
move denoiser to above SpacegroupAnalyzer
since we want to make sure the denoising happens before doing a symmetry check. Though, it probably isn't a big deal whether you pass a conventional vs. primitive cell to the relaxer https://github.com/materialsvirtuallab/m3gnet/issues/10
