### 2022-06-17 95f2e81
hopefully renaming to S doesn't undo the fix for inefficient memory usage
