### 2022-06-17 90d8546
change most tests to relax_on_decode=False since True is now default behavior
