### 2022-06-17 47c0710
remove (erroneous) typehint and doc for xtal2png supporting directory paths
i.e. you should pass a list of paths or a list of structures, not a directory containing relevant files
