### 2022-06-17 04e9abc
Create logging_filter.py
Can remove after next m3gnet release
https://github.com/materialsvirtuallab/m3gnet/issues/11
