+++ src/xtal2png/utils/logging_filter.py
@@ -1,0 +1,31 @@
+import re
+
+
+class Filter(object):
+    """https://stackoverflow.com/a/63662744/13697228
+
+    For reassigning default, https://stackoverflow.com/a/51340381/13697228"""
+
+    def __init__(self, stream, re_pattern):
+        self.stream = stream
+        self.pattern = (
+            re.compile(re_pattern) if isinstance(re_pattern, str) else re_pattern
+        )
+        self.triggered = False
+
+    def __getattr__(self, attr_name):
+        return getattr(self.stream, attr_name)
+
+    def write(self, data):
+        if data == "\n" and self.triggered:
+            self.triggered = False
+        else:
+            if self.pattern.search(data) is None:
+                self.stream.write(data)
+                self.stream.flush()
+            else:
+                # caught bad pattern
+                self.triggered = True
+
+    def flush(self):
+        self.stream.flush()
