### 2022-06-17 25ab4d9
increase tol_multiplier from 1 to 4 for the relaxation test
