### 2022-06-17 11ed56f
Revert "Create logging_filter.py"
This reverts commit 04e9abc7a713483f3f48ef8ce65a71ff72f7a586.
