### 2022-06-17 1eadcc1
m3gnet 0.0.6 enables verbosity control
https://github.com/materialsvirtuallab/m3gnet/issues/11#issuecomment-1158445700
