### 2022-06-17 2afd856
update README with m3gnet usage
