### 2022-06-17 36af5f7
Create imagen_pytorch_example.py
