### 2022-06-17 4bda7a8
update imagen_pytorch_example.py with xtal2png usage
