### 2022-06-17 5bb0498
structures <---> arrays functions get rgb_output kwarg, default True
