+++ src/xtal2png/core.py
@@ -24,7 +24,13 @@ from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from tqdm import tqdm
 
 from xtal2png import __version__
-from xtal2png.utils.data import dummy_structures, rgb_scaler, rgb_unscaler
+from xtal2png.utils.data import (
+    dummy_structures,
+    element_wise_scaler,
+    element_wise_unscaler,
+    rgb_scaler,
+    rgb_unscaler,
+)
 
 # from sklearn.preprocessing import MinMaxScaler
 
@@ -473,6 +479,7 @@ class XtalConverter:
     def structures_to_arrays(
         self,
         structures: Sequence[Structure],
+        rgb_output=True,
     ):
         """Convert pymatgen Structure to scaled 3D array of crystallographic info.
 
@@ -580,24 +587,62 @@ class XtalConverter:
         frac_coords = np.stack(frac_coords_tmp)
         distance_matrix = np.stack(distance_matrix_tmp)
 
-        # REVIEW: consider using modified pettifor scale instead of atomic numbers
-        # REVIEW: consider using feature_range=atom_range or 2*atom_range
-        # REVIEW: since it introduces a sort of non-linearity b.c. of rounding
-        atom_scaled = rgb_scaler(atomic_numbers, data_range=self.atom_range)
-        frac_scaled = rgb_scaler(frac_coords, data_range=self.frac_range)
-        a_scaled = rgb_scaler(latt_a, data_range=self.a_range)
-        b_scaled = rgb_scaler(latt_b, data_range=self.b_range)
-        c_scaled = rgb_scaler(latt_c, data_range=self.c_range)
-        angles_scaled = rgb_scaler(angles, data_range=self.angles_range)
-        volume_scaled = rgb_scaler(volume, data_range=self.volume_range)
-        space_group_scaled = rgb_scaler(space_group, data_range=self.space_group_range)
-        # NOTE: max_distance could be added as another (repeated) value/row to infer
-        # NOTE: kind of like frac_distance_matrix, not sure if would be effective
-        # NOTE: Or could normalize distance_matix by cell volume
-        # NOTE: and possibly include cell volume as a (repeated) value/row to infer
-        # NOTE: It's possible extra info like this isn't so bad, instilling the physics
-        # NOTE: but it could also just be extraneous work to predict/infer
-        distance_scaled = rgb_scaler(distance_matrix, data_range=self.distance_range)
+        if rgb_output:
+            # REVIEW: consider using modified pettifor scale instead of atomic numbers
+            # REVIEW: consider using feature_range=atom_range or 2*atom_range
+            # REVIEW: since it introduces a sort of non-linearity b.c. of rounding
+            atom_scaled = rgb_scaler(atomic_numbers, data_range=self.atom_range)
+            frac_scaled = rgb_scaler(frac_coords, data_range=self.frac_range)
+            a_scaled = rgb_scaler(latt_a, data_range=self.a_range)
+            b_scaled = rgb_scaler(latt_b, data_range=self.b_range)
+            c_scaled = rgb_scaler(latt_c, data_range=self.c_range)
+            angles_scaled = rgb_scaler(angles, data_range=self.angles_range)
+            volume_scaled = rgb_scaler(volume, data_range=self.volume_range)
+            space_group_scaled = rgb_scaler(
+                space_group, data_range=self.space_group_range
+            )
+            # NOTE: max_distance could be added as another (repeated) value/row to infer
+            # NOTE: kind of like frac_distance_matrix, not sure if would be effective
+            # NOTE: Or could normalize distance_matix by cell volume
+            # NOTE: and possibly include cell volume as a (repeated) value/row to infer
[…]
