### 2022-06-17 a06542f
rename rgb_output to rgb_scaling for clarity, fix by adding missing feature_range=
