### 2022-06-17 b33a42f
add tests for rgb_scaling=False
