### 2022-06-17 04ba552
refactor to use rgb_scaling=False
