+++ scripts/imagen_pytorch_example.py
@@ -1,12 +1,10 @@
-from os import path
-from pathlib import Path
-from uuid import uuid4
-
-import torch
-from imagen_pytorch import Imagen, ImagenTrainer, Unet
+from imagen_pytorch import Imagen, ImagenTrainer, SRUnet256, Unet
 from mp_time_split.core import MPTimeSplit
 
 from xtal2png.core import XtalConverter
+
+low_mem = True
+max_batch_size = 32
 
 mpt = MPTimeSplit()
 mpt.load()
@@ -14,16 +12,8 @@ mpt.load()
 fold = 0
 train_inputs, val_inputs, train_outputs, val_outputs = mpt.get_train_and_val_data(fold)
 
-data_path = path.join("data", "preprocessed", "mp-time-split", f"fold={fold}")
-xc = XtalConverter(
-    save_dir=data_path, encode_as_primitive=True, decode_as_primitive=True
-)
-xc.xtal2png(train_inputs.tolist())
-
-max_batch_size = 32
-
-results_folder = path.join("data", "interim", "ddpm", f"fold={fold}", str(uuid4())[0:4])
-Path(results_folder).mkdir(exist_ok=True, parents=True)
+xc = XtalConverter(save_dir="tmp", encode_as_primitive=True, decode_as_primitive=True)
+training_images = xc.structures_to_arrays(train_inputs.tolist(), rgb_scaling=False)
 
 # unets for unconditional imagen
 
@@ -36,14 +26,17 @@ unet1 = Unet(
     use_linear_attn=True,
 )
 
-unet2 = Unet(
-    dim=32,
-    dim_mults=(1, 2, 4),
-    num_resnet_blocks=3,
-    layer_attns=(False, True, True),
-    layer_cross_attns=(False, True, True),
-    use_linear_attn=True,
-)
+if low_mem:
+    unet2 = Unet(
+        dim=32,
+        dim_mults=(1, 2, 4),
+        num_resnet_blocks=3,
+        layer_attns=(False, True, True),
+        layer_cross_attns=(False, True, True),
+        use_linear_attn=True,
+    )
+else:
+    unet2 = SRUnet256()
 
 # imagen, which contains the unets above (base unet and super resoluting ones)
 
@@ -51,15 +44,11 @@ imagen = Imagen(
     condition_on_text=False,  # this must be set to False for unconditional Imagen
     unets=(unet1, unet2),
     channels=1,
-    image_sizes=(16, 32),
+    image_sizes=(32, 64),
     timesteps=1000,
 )
 
 trainer = ImagenTrainer(imagen).cuda()
-
-# now get a ton of images and feed it through the Imagen trainer
-
-training_images = torch.randn(4, 1, 64, 64).cuda()
 
 # train each unet in concert, or separately (recommended) to completion
 
