### 2022-06-18 be36bdd
Update imagen_pytorch_example.py
