### 2022-06-18 cc612e0
Update imagen_pytorch_example.py
