### 2022-06-18 0ad5087
try except for m3gnet import if relax_on_decode
