### 2022-06-18 c95277a
write get_image_mode function
