### 2022-06-18 19851ae
support RGB (3-channel) averaging in addition to grayscale images
