### 2022-06-18 47fea9b
use get_image_mode
