### 2022-06-18 e200a10
three channel unit tests
