### 2022-06-18 21a9bdd
apply appropriate transpose function when loading images
