### 2022-06-18 19548de
ensure that self.channels is at least 1
