### 2022-06-18 c30f7a3
remove check to avoid strange mypy warning
Seems unlikely that people would pass 0 or a negative value to channels anyway
