+++ src/xtal2png/core.py
@@ -20,6 +20,9 @@ from pymatgen.core.structure import Structure
 from pymatgen.io.cif import CifWriter
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from tqdm import tqdm
+from element_coder.encoder import encode
+from element_coder.decoder import decode
+from element_coder.utils import get_range
 
 from xtal2png import __version__
 from xtal2png.utils.data import (
@@ -30,6 +33,8 @@ from xtal2png.utils.data import (
     rgb_scaler,
     rgb_unscaler,
 )
+
+from functools import lru_cache
 
 # from sklearn.preprocessing import MinMaxScaler
 
@@ -71,6 +76,11 @@ DISTANCE_KEY = "distance"
 def construct_save_name(s: Structure) -> str:
     save_name = f"{s.formula.replace(' ', '')},volume={int(np.round(s.volume))},uid={str(uuid4())[0:4]}"  # noqa: E501
     return save_name
+
+
+@lru_cache(maxsize=None)
+def _element_encoding_range_cached(elements, encoding_type):
+    return get_range(elements, encoding_type)
 
 
 class XtalConverter:
@@ -143,6 +153,10 @@ class XtalConverter:
         func:``XtalConverter().arrays_to_structures`` directly instead.
     verbose: bool, optional
         Whether to print verbose debugging information or not.
+    element_encoding : str, optional
+        How to encode the element. Can be one of `element_coder.data.coding_data._PROPERTY_KEYS`
+        (e.g., `mod_pettifor`, `atomic`, `pettifor`, `X`). Defaults to `atomic`
+        (which encodes elements as atomic numbers).
 
     Examples
     --------
@@ -171,6 +185,7 @@ class XtalConverter:
         relax_on_decode: bool = True,
         channels: int = 1,
         verbose: bool = True,
+        element_encoding: Optional[str] = "atomic",
     ):
         """Instantiate an XtalConverter object with desired ranges and ``max_sites``."""
         self.atom_range = atom_range
@@ -184,6 +199,7 @@ class XtalConverter:
         self.distance_range = distance_range
         self.max_sites = max_sites
         self.save_dir = save_dir
+        self.element_encoding = element_encoding
 
         if isinstance(symprec, (float, int)):
             self.encode_symprec = symprec
@@ -205,8 +221,14 @@ class XtalConverter:
 
         self.channels = channels
         self.verbose = verbose
-
         Path(save_dir).mkdir(exist_ok=True, parents=True)
+
+    @property
+    def _element_encoding_range(self):
+        # We do *not* use cached_property as the result might change as
+        # the users calls fit. However, we still want to cache, as we reuse
+        # the result of this method for both encoding and decoding.
+        return _element_encoding_range_cached(self.atom_range, self.element_encoding)
 
     def xtal2png(
         self,
@@ -543,7 +565,7 @@ class XtalConverter:
             raise ValueError("`structures` should be a list of pymatgen Structure(s)")
 
         # extract crystallographic information
-        atomic_numbers: List[List[int]] = []
+        element_encoding: List[List[int]] = []
         frac_coords_tmp: List[NDArray] = []
         latt_a: List[float] = []
         latt_b: List[float] = []
@@ -574,9 +596,12 @@ class XtalConverter:
                 raise ValueError(
                     f"crystal supplied with {n_sites} sites, which is more than {self.max_sites} sites. Remove crystal or increase `max_sites`."  # noqa
                 )
-            atomic_numbers.append(
+            element_encoding.append(
                 np.pad(
-                    list(s.atomic_numbers),
+                    [
+                        encode(atomic_number, self.element_encoding)
+                        for atomic_number in s.atomic_numbers
+                    ],
                     (0, self.max_sites - n_sites),
                 ).tolist()
[…]
