+++ setup.cfg
@@ -55,6 +55,7 @@ install_requires =
     plotly
     kaleido
     m3gnet
+    element-coder
 
 
 [options.packages.find]
