### 2022-06-20 7b2a10c
remove leftover sys.stdout restoration
