### 2022-06-20 8d8eb57
remove extraneous expand_dims
