+++ setup.cfg
@@ -54,7 +54,6 @@ install_requires =
     pymatgen
     plotly
     kaleido
-    m3gnet
 
 
 [options.packages.find]
@@ -66,6 +65,7 @@ exclude =
 # Add here additional requirements for extra features, to install with:
 # `pip install xtal2png[PDF]` like:
 # PDF = ReportLab; RXP
+relax = m3gnet
 
 # Add here test requirements (semicolon/line-separated)
 testing =
