### 2022-06-20 efdebfd
add m3gnet as test requirement
