+++ setup.cfg
@@ -72,6 +72,7 @@ testing =
     setuptools
     pytest
     pytest-cov
+    m3gnet
 
 [options.entry_points]
 # Add here console scripts like:
