### 2022-06-22 c271125
docs(CHANGELOG): update release notes
