+++ CHANGELOG.md
@@ -1,12 +1,82 @@
 # CHANGELOG
 
-## Unreleased
+## [v0.6.2](https://github.com/sparks-baird/xtal2png/releases/tag/v0.6.2) - 2022-06-20 18:47:21
 
-Changes unreleased.
+<!-- Release notes generated using configuration in .github/release.yml at v0.6.2 -->
+
+## What's Changed
+### Other Changes
+* try except for m3gnet import if relax_on_decode by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/110
+* refactor to use rgb_scaling=False in imagen-pytorch example by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/107
+* support RGB (3-channel) averaging in addition to grayscale images by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/111
+* add num_sites and a note about displaying plotly figures by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/115
+* flat is better than nested (fit refactor) by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/113
+* commented "shuffle" code (might test later) and add channels variable to example scripts by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/112
+* make m3gnet an optional dep and relax_on_decode False by default by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/118
+* tqdm_if_verbose wrapper fn by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/120
+
+
+**Full Changelog**: https://github.com/sparks-baird/xtal2png/compare/v0.5.1...v0.6.2
+
+### Refactor
+
+- general:
+  - refactor to use rgb_scaling=False ([04ba552](https://github.com/sparks-baird/xtal2png/commit/04ba5525646dc2c7afa5e22e90855ea9998b9a4d)) ([#107](https://github.com/sparks-baird/xtal2png/pull/107))
+
+## [v0.5.1](https://github.com/sparks-baird/xtal2png/releases/tag/v0.5.1) - 2022-06-18 02:40:01
+
+<!-- Release notes generated using configuration in .github/release.yml at v0.5.1 -->
+
+## What's Changed
+### Other Changes
+* structures <---> arrays functions get rgb_output kwarg, default True by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/104
+* imagen-pytorch example by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/106
+* Add tests for rgb_scaling=False by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/105
+
+
+**Full Changelog**: https://github.com/sparks-baird/xtal2png/compare/v0.5.0...v0.5.1
+
+## [v0.5.0](https://github.com/sparks-baird/xtal2png/releases/tag/v0.5.0) - 2022-06-17 04:13:38
+
+<!-- Release notes generated using configuration in .github/release.yml at v0.5.0 -->
+
+## What's Changed
+* ❗Implement [`m3gnet`](https://github.com/materialsvirtuallab/m3gnet)'s [DFT surrogate structure relaxation](https://github.com/materialsvirtuallab/m3gnet#structure-relaxation) during decoding and use by default by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/100.
+  * (Much thanks to @shyuep for [support on this](https://github.com/materialsvirtuallab/m3gnet/issues?q=is%3Aissue+commenter%3Asgbaird))
+* fit method for feature ranges by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/91 as suggested by @kjappelbaum
[…]
