### 2022-06-22 2d78104
add badges and other image-to-image models
