+++ docs/index.md
@@ -1,8 +1,22 @@
 # xtal2png
 
-Encode/decode a crystal structure to/from a grayscale PNG image for direct use with image-based machine learning models such as [Palette].
+Encode/decode a crystal structure to/from a grayscale PNG image for direct use with image-based machine learning models such as [Imagen], [DALLE2], or [Palette].[^1]
 
-[![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sparks-baird/xtal2png/blob/main/notebooks/1.0-xtal2png-tutorial.ipynb)
+<!-- [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sparks-baird/xtal2png/blob/main/notebooks/1.0-xtal2png-tutorial.ipynb) [![status](https://joss.theoj.org/papers/0c704f6ae9739c1e97e05ae0ad57aecb/status.svg)](https://github.com/openjournals/joss-papers/blob/joss.04465/joss.04465/10.21105.joss.04465.pdf) https://joss.theoj.org/papers/0c704f6ae9739c1e97e05ae0ad57aecb -->
+
+ <a class="github-button" href="https://github.com/sparks-baird/xtal2png"
+ data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star
+ sparks-baird/xtal2png on GitHub">Star</a>
+ <a class="github-button"
+ href="https://github.com/sgbaird" data-size="large" data-show-count="true"
+ aria-label="Follow @sgbaird on GitHub">Follow @sgbaird</a>
+ <a class="github-button" href="https://github.com/sparks-baird/xtal2png/issues"
+ data-icon="octicon-issue-opened" data-size="large" data-show-count="true"
+ aria-label="Issue sparks-baird/xtal2png on GitHub">Issue</a>
+ <a class="github-button" href="https://github.com/sparks-baird/xtal2png/discussions" data-icon="octicon-comment-discussion" data-size="large" aria-label="Discuss sparks-baird/xtal2png on GitHub">Discuss</a>
+ <br><br>
+ 
+ <!-- ![GitHub Repo stars](https://img.shields.io/github/stars/sparks-baird/xtal2png?style=social) ![GitHub followers](https://img.shields.io/github/followers/sgbaird?style=social) ![GitHub issues](https://img.shields.io/github/issues-raw/sparks-baird/xtal2png) ![GitHub closed issues](https://img.shields.io/github/issues-closed-raw/sparks-baird/xtal2png) -->
 
 ## Contents
 
@@ -28,3 +42,9 @@ Module Reference <api/modules>
 [reStructuredText]: http://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html
 [MyST]: https://myst-parser.readthedocs.io/en/latest/
 [Palette]: https://iterative-refinement.github.io/palette/
+[Janspiry/Palette-Image-to-Image-Diffusion-Models]: https://github.com/Janspiry/Palette-Image-to-Image-Diffusion-Models
+[Imagen]: https://imagen.research.google/
+[lucidrains/imagen-pytorch]: https://github.com/lucidrains/imagen-pytorch#usage
+[DALLE2]: https://openai.com/dall-e-2/
+[lucidrains/DALLE2-pytorch]: https://github.com/lucidrains/DALLE2-pytorch#unconditional-training
+[^1]: For unofficial implementations, see [lucidrains/imagen-pytorch], [lucidrains/DALLE2-pytorch], and [Janspiry/Palette-Image-to-Image-Diffusion-Models], respectively
