### 2022-06-22 1fcf537
fixup badges in index.md
