+++ docs/index.md
@@ -2,19 +2,26 @@
 
 Encode/decode a crystal structure to/from a grayscale PNG image for direct use with image-based machine learning models such as [Imagen], [DALLE2], or [Palette].[^1]
 
-<!-- [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sparks-baird/xtal2png/blob/main/notebooks/1.0-xtal2png-tutorial.ipynb) [![status](https://joss.theoj.org/papers/0c704f6ae9739c1e97e05ae0ad57aecb/status.svg)](https://github.com/openjournals/joss-papers/blob/joss.04465/joss.04465/10.21105.joss.04465.pdf) https://joss.theoj.org/papers/0c704f6ae9739c1e97e05ae0ad57aecb -->
+[![Open In
+Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sparks-baird/xtal2png/blob/main/notebooks/1.0-xtal2png-tutorial.ipynb)
+[![status](https://joss.theoj.org/papers/0c704f6ae9739c1e97e05ae0ad57aecb/status.svg)](https://github.com/openjournals/joss-papers/blob/joss.04465/joss.04465/10.21105.joss.04465.pdf)
+[![PyPI -
+Downloads](https://img.shields.io/pypi/dm/xtal2png)](https://pypi.org/project/xtal2png)
+[![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/xtal2png?style=flat&color=blue&label=Conda%20Downloads)](https://anaconda.org/conda-forge/xtal2png)
 
- <a class="github-button" href="https://github.com/sparks-baird/xtal2png"
- data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star
- sparks-baird/xtal2png on GitHub">Star</a>
- <a class="github-button"
- href="https://github.com/sgbaird" data-size="large" data-show-count="true"
- aria-label="Follow @sgbaird on GitHub">Follow @sgbaird</a>
- <a class="github-button" href="https://github.com/sparks-baird/xtal2png/issues"
- data-icon="octicon-issue-opened" data-size="large" data-show-count="true"
- aria-label="Issue sparks-baird/xtal2png on GitHub">Issue</a>
- <a class="github-button" href="https://github.com/sparks-baird/xtal2png/discussions" data-icon="octicon-comment-discussion" data-size="large" aria-label="Discuss sparks-baird/xtal2png on GitHub">Discuss</a>
- <br><br>
+<a class="github-button" href="https://github.com/sparks-baird/xtal2png"
+data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star
+sparks-baird/xtal2png on GitHub">Star</a>
+<a class="github-button"
+href="https://github.com/sgbaird" data-size="large" data-show-count="true"
+aria-label="Follow @sgbaird on GitHub">Follow @sgbaird</a>
+<a class="github-button" href="https://github.com/sparks-baird/xtal2png/issues"
+data-icon="octicon-issue-opened" data-size="large" data-show-count="true"
+aria-label="Issue sparks-baird/xtal2png on GitHub">Issue</a>
+<a class="github-button" href="https://github.com/sparks-baird/xtal2png/discussions" data-icon="octicon-comment-discussion" data-size="large" aria-label="Discuss sparks-baird/xtal2png on GitHub">Discuss</a>
+<br><br>
+
+> Manuscript coming soon! (est. July 22, 2022)
 
[…]
