### 2022-06-23 cbba635
encode/decode as primitive True by default (temporary fix)
