+++ src/xtal2png/core.py
@@ -166,8 +166,8 @@ class XtalConverter:
         save_dir: Union[str, "PathLike[str]"] = path.join("data", "preprocessed"),
         symprec: Union[float, Tuple[float, float]] = 0.1,
         angle_tolerance: Union[float, int, Tuple[float, float], Tuple[int, int]] = 5.0,
-        encode_as_primitive: bool = False,
-        decode_as_primitive: bool = False,
+        encode_as_primitive: bool = True,
+        decode_as_primitive: bool = True,
         relax_on_decode: bool = False,
         channels: int = 1,
         verbose: bool = True,
