### 2022-06-23 1653229
refactor multiple for loops into single for loop and enhance mem efficiency
Thanks to @faris-k for pointing out the behavior while running it
