### 2022-06-23 d7963fa
refactor to unit_cell_converter function
