### 2022-06-23 523db7e
use encode/decode cell_type-s instead of as_primitive
