### 2022-06-23 e0a4075
update kwargs in tests
