### 2022-06-23 b2b0403
get_image_mode docstring
