### 2022-06-24 c5d09fb
mask_types and fit docstring
