### 2022-06-24 a2f1e24
not isinstance(images, list) ValueError
