### 2022-06-24 7537dff
masking based on information type
