### 2022-06-24 7a26fb1
test for distance mask
Consider implementing mask tests for other mask types as well
