### 2022-06-24 f3e9541
rename "latt_a" to "a" ("b", "c") for consistency with kwarg naming
