### 2022-06-24 6488685
test_mask_error and fix test_distance_mask
