### 2022-06-24 3d24940
allow for "lower_tri" mask type
