### 2022-06-24 0853d8b
test_lower_tri_mask
