### 2022-06-24 6228ebe
max_sites error msg: note about encode/decode_cell_type kwargs
