### 2022-06-24 72d7f1c
normalize lengths by the cubed root of the number of atoms
