### 2022-06-24 837ff2e
element_wise_scalers, rgb_scalers, etc. (looped versions)
