### 2022-06-24 3e6ac7f
test_normalize_lengths WIP
