+++ setup.cfg
@@ -54,7 +54,7 @@ install_requires =
     pymatgen
     plotly
     kaleido
-    element-coder
+    element_coder
 
 
 [options.packages.find]
