### 2022-06-25 6dbfd34
chore: use decode_many
