### 2022-06-25 35c7ef9
chore: expose decode metric
