+++ setup.cfg
@@ -54,7 +54,7 @@ install_requires =
     pymatgen
     plotly
     kaleido
-    element_coder
+    element-coder
 
 
 [options.packages.find]
