### 2022-06-25 08ac35a
fixup getting started instructions, linting, mention m3gnet as optional
