### 2022-06-28 3bfa9fe
keep a _atom_range variable that will probably get removed later
        uniq_atoms = np.unique(list(chain(*atomic_numbers)))
        self._atom_range = [np.min(uniq_atoms), np.max(uniq_atoms)]
