### 2022-06-28 918a42c
check against _atom_range instead of atom_range during test_fit
