+++ src/xtal2png/core.py
@@ -58,7 +58,7 @@ A_ID = 3
 B_ID = 4
 C_ID = 5
 ANGLES_ID = 6
-VOLUME_ID = 7
+NUM_SITES_ID = 7
 SPACE_GROUP_ID = 8
 DISTANCE_ID = 9
 
@@ -68,7 +68,7 @@ A_KEY = "a"
 B_KEY = "b"
 C_KEY = "c"
 ANGLES_KEY = "angles"
-VOLUME_KEY = "volume"
+NUM_SITES_KEY = "num_sites"
 SPACE_GROUP_KEY = "space_group"
 DISTANCE_KEY = "distance"
 LOWER_TRI_KEY = "lower_tri"
@@ -80,7 +80,7 @@ SUPPORTED_MASK_KEYS = [
     B_KEY,
     C_KEY,
     ANGLES_KEY,
-    VOLUME_KEY,
+    NUM_SITES_KEY,
     SPACE_GROUP_KEY,
     DISTANCE_KEY,
     LOWER_TRI_KEY,
@@ -88,8 +88,8 @@ SUPPORTED_MASK_KEYS = [
 
 
 def construct_save_name(s: Structure) -> str:
-    """Construct savename based on formula, volume, and a uid."""
-    save_name = f"{s.formula.replace(' ', '')},volume={int(np.round(s.volume))},uid={str(uuid4())[0:4]}"  # noqa: E501
+    """Construct savename based on formula, space group, and a uid."""
+    save_name = f"{s.formula.replace(' ', '')},space-group={int(np.round(s.get_space_group_info()[1]))},uid={str(uuid4())[0:4]}"  # noqa: E501
     return save_name
 
 
@@ -128,8 +128,8 @@ class XtalConverter:
         Expected range for lattice parameter length c, by default (2.0, 36.0)
     angles_range : Tuple[float, float], optional
         Expected range for lattice parameter angles, by default (0.0, 180.0)
-    volume_range : Tuple[float, float], optional
-        Expected range for unit cell volumes, by default (0.0, 1000.0)
+    num_sites_range : Tuple[float, float], optional
+        Expected range for unit cell num_sites, by default (0, 52)
     space_group_range : Tuple[int, int], optional
         Expected range for space group numbers, by default (1, 230)
     distance_range : Tuple[float, float], optional
@@ -185,7 +185,7 @@ class XtalConverter:
         'sokalsneath', 'sqeuclidean', 'yule'. Defaults to "euclidean".
     mask_types : List[str], optional
         List of information types to mask out (assign as 0) from the array/image. values
-        are "atom", "frac", "a", "b", "c", "angles", "volume", "space_group",
+        are "atom", "frac", "a", "b", "c", "angles", "num_sites", "space_group",
         "distance", "diagonal", and None. If None, then no masking is applied. If
         "diagonal" is present, then zeros out the lower triangle. By default, None.
 
@@ -204,7 +204,7 @@ class XtalConverter:
         b_range: Tuple[float, float] = (2.0, 15.0),
         c_range: Tuple[float, float] = (2.0, 36.0),
         angles_range: Tuple[float, float] = (0.0, 180.0),
-        volume_range: Tuple[float, float] = (0.0, 1500.0),
+        num_sites_range: Tuple[float, float] = (0, 52),
         space_group_range: Tuple[int, int] = (1, 230),
         distance_range: Tuple[float, float] = (0.0, 18.0),
         max_sites: int = 52,
@@ -227,7 +227,7 @@ class XtalConverter:
         self.b_range = b_range
         self.c_range = c_range
         self.angles_range = angles_range
-        self.volume_range = volume_range
+        self.num_sites_range = num_sites_range
         self.space_group_range = space_group_range
[…]
