### 2022-06-29 294fe01
apply num_sites masking during encode and decode
