+++ src/xtal2png/core.py
@@ -394,7 +394,7 @@ class XtalConverter:
         b = [s.lattice.b for s in S]
         c = [s.lattice.c for s in S]
         space_group = [s.get_space_group_info()[1] for s in S]
-        num_sites = [len(s.sites) for s in S]
+        num_sites = [s.num_sites for s in S]
         distance = [s.distance_matrix for s in S]
 
         if verbose:
@@ -525,7 +525,7 @@ class XtalConverter:
 
     def png2xtal(
         self, images: List[Union[Image.Image, "PathLike"]], save: bool = False
-    ):
+    ) -> List[Structure]:
         """Decode PNG files as Structure objects.
 
         Parameters
@@ -673,7 +673,7 @@ class XtalConverter:
             latt_b.append(s._lattice.b)
             latt_c.append(s._lattice.c)
             angles.append(list(s._lattice.angles))
-            num_sites.append(len(s.sites))
+            num_sites.append(s.num_sites)
             space_group.append(s.get_space_group_info()[1])
 
             if n_sites != s.distance_matrix.shape[0]:
@@ -806,6 +806,10 @@ class XtalConverter:
         ]
         id_data = self.assemble_blocks(*id_blocks)
 
+        # apply num_sites mask (zero out bottom and RHS blocks past num_sites)
+        data = self.apply_num_sites_mask(data, num_sites)
+        id_data = self.apply_num_sites_mask(id_data, num_sites)
+
         data = np.expand_dims(data, 1)
         id_data = np.expand_dims(id_data, 1)
 
@@ -885,20 +889,23 @@ class XtalConverter:
         return data
 
     def disassemble_blocks(
-        self, data, id_data: Optional[NDArray] = None, id_mapper: Optional[dict] = None
+        self,
+        data,
+        # id_data: Optional[NDArray] = None,
+        # id_mapper: Optional[dict] = None,
     ):
-        if (id_data is None) is not (id_mapper is None):
-            raise ValueError(
-                f"id_data (type: {type(id_data)}) and id_mapper (type: {type(id_mapper)}) should either both be assigned or both be None."  # noqa
-            )
-        elif id_data is None and id_mapper is None:
-            _, id_data, id_mapper = self.structures_to_arrays(dummy_structures)
+        # TODO: implement a more robust solution using id_data and id_mapper
 
-        assert (
-            id_data is not None and id_mapper is not None
-        ), "id_data and id_mapper should not be None at this point"
+        # if (id_data is None) is not (id_mapper is None):
+        #     raise ValueError(
+        #         f"id_data (type: {type(id_data)}) and id_mapper (type: {type(id_mapper)}) should either both be assigned or both be None."  # noqa
+        #     )
+        # elif id_data is None and id_mapper is None:
+        #     _, id_data, id_mapper = self.structures_to_arrays(dummy_structures)
 
-        [a.shape for a in np.array_split(data, [12], axis=1)]
+        # assert (
+        #     id_data is not None and id_mapper is not None
+        # ), "id_data and id_mapper should not be None at this point"
 
         zero_pad = 12
         left_arr, right_arr = np.array_split(data, [zero_pad], axis=1)
@@ -974,9 +981,38 @@ class XtalConverter:
 
         # convert to single channel and remove singleton dimension before disassembly
         data = np.mean(data, axis=1)
-        if id_data is not None:
-            id_data = np.mean(id_data, axis=1)
-        arrays = self.disassemble_blocks(data, id_data=id_data, id_mapper=id_mapper)
+
+        # to extract num_sites for preprocess masking of data
[…]
