### 2022-06-29 a399bf4
test_num_sites_mask and use num_sites instead of atoms for test_mask_error
