### 2022-06-29 #4528 @editorialbot
[REVIEW]: xtal2png: A Python package for representing crystal structure as PNG files
**Submitting author:** @sgbaird (<a href="http://orcid.org/0000-0002-4491-6876">Sterling Baird</a>)
**Repository:** https://github.com/sparks-baird/xtal2png
**Branch with paper.md** (empty if default branch): main
**Version:** 0.9.4
