### 2022-07-07 sparks-baird/xtal2png#144 @PeterKraus
JOSS paper review - Software paper
In this issue, I will collate points that relate to the software paper itself. Link to review is: https://github.com/openjournals/joss-reviews/issues/4528

- [x] the reference to the software repo seems unnecessary here (and throughout the paper). The link to the repo is right there in the front-matter!
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L45

- [x] the summary is not clear about what is being encoded into what. My understanding is that `xtal2png` can take a coordinate file and create a fingerprint / QR-code-like representation of that coordinate file. If that's the case, it should be more explicit in the summary.
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L46

- [x] streamlined results? Is the point here that it feeds well into image-based pipelines?
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L47

- [x] The references in this section on domain transfer make it hard to read. Please rewrite - perhaps put the citations behind the dates.
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L64-L67

- [x] Here, rather than to-from, I'd say between coordinate file and png representation (or something similar):
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L73-L74

- [x] I am not sure I understand the relevance of Pallette. Please elaborate.

- [x] It would be helpful to discuss the technical limitations somewhere. Perhaps it's in the docs, but in my view it could be commented on in the paper:
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L94-L97
- First, is there a reason 64x64 pixel image has to be used? I understand it's possible to encode at most 52 atoms within the unit cell. Just by reading the paper, it's not obvious why the cell parameters (lattice constants, angles, volume, point group) occupy rows - seems like a waste of space. If that information was moved into the upper-left zero sector, there would be more space for the pairwise distance matrix.
- Second, the issue of 255 possible values is a limitation of the grayscale requirement. If this was lifted, the precision could go up dramatically.
- Third, I'm sure there's a good reason for it, but what's the distance matrix for? It can be calculated from the fractional coordinates. If all we want is to do is to represent the CIF file, there's no reason to include it.

- [x] I would rewrite this section, as a "potential" time saving is only potential.
https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/reports/paper.md?plain=1#L105-L106

- [x] There is no mention of any other software that provides similar functionality. If there truly is not any encoder/decoder for crystal structure to image data, perhaps a short section discussing existing software for molecules and/or textual representation could be included.
