### 2022-07-07 73f725e
remove links to xtal2png repo (already in frontmatter)
https://github.com/sparks-baird/xtal2png/issues/144#issue-1296952657

> the reference to the software repo seems unnecessary here (and throughout the paper). The link to the repo is right there in the front-matter!
