### 2022-07-07 22bf477
clarify encoding/decoding in summary and provide context for "streamlined"
https://github.com/sparks-baird/xtal2png/issues/144#issue-1296952657
