### 2022-07-07 sparks-baird/xtal2png#146 @PeterKraus
JOSS paper review - Documentation
Here I will collate issues/suggestions I found with documentation. See https://github.com/openjournals/joss-reviews/issues/4528 for the full JOSS review.

- [x] Statement of need - the landing page on GitHub has a nice `README.md`, but the index page of the docs could be expanded with a short blurb. Feel free to reuse something from the paper and don't be afraid of being more technical.
- [x] Functionality documentation - the API docs are built, which is great, but I'm missing a "functionality" page that would link to classes/functions in the API docs. It's up to you if you put it as a subsection of the "Overview" or make a new section. I think such a page should discuss the decisions behind the default output format, link to the key classes/functions in the API docs, and also mention it's possible to customise the output PNGs. Then, reference this overview from the paper to avoid having to write the same thing twice and risking the paper might end up out of date.
- [x] Automated tests - CI is enabled and tests pass. However, it's difficult to figure out what's being tested where. I'd suggest splitting [this](https://github.com/sparks-baird/xtal2png/blob/cc644f3b746bc4d106a273ad85f8b45df62edf4a/tests/xtal2png_test.py) test file into multiple modules, separating the functionality tests from the validity tests. Also, I think a very useful category of tests is missing - comparing image data directly with a reference png. You can use the instructions [here](https://stackoverflow.com/questions/35176639/compare-images-python-pil) to get a test working. We have written a similar test you can copy, [here](https://github.com/dgbowl/dgpost/blob/913570adb00c6bc4afaffb3d6495d665c2d37e70/tests/utils.py#L92-L117).
