### 2022-07-07 #4528 @PeterKraus
Alright, this is quite an interesting software package. I've made a couple of issues on the project github, I will play around with the software over the next few days. @sgbaird, please let me know (here or in the issues) once you want me to have another look.
