### 2022-07-07 d98f8fd
add motivation and somewhat technical blurb, remove estimated date of manuscript
