### 2022-07-07 d833f5c
explanation of Palette, design limitations link and comment, refactor 2nd-to-last paragraph
