+++ reports/paper.md
@@ -11,6 +11,9 @@ authors:
     orcid: 0000-0002-4491-6876
     equal-contrib: false
     affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: Kevin M. Jablonka
+    orcid: 0000-0003-4894-4660
+    affiliation: "3"
   - name: Hasan M. Sayeed
     orcid: 0000-0002-6583-7755
     equal-contrib: false
@@ -19,6 +22,9 @@ authors:
     orcid: 0000-0002-4857-7584
     equal-contrib: false
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Colton Seegmiller
+  - equal-contrib: false
+  - affiliation: "4" # (Multiple affiliations must be quoted)
   - name: Taylor D. Sparks
     orcid: 0000-0001-8020-7711
     equal-contrib: false
@@ -28,6 +34,10 @@ affiliations:
    index: 1
  - name: Computer Science, University of Southern California, USA
    index: 2
+ - name: École Polytechnique Fédérale de Lausanne, Switzerland
+   index: 3
+ - name: Computer Science, Utah Valley University, USA
+   index: 4
 date: 9 June 2022
 bibliography: paper.bib
 
