+++ docs/conf.py
@@ -76,6 +76,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx_rtd_theme",
     "sphinx_copybutton",
+    "nbsphinx",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
