### 2022-07-07 8f61d5e
docs: only mention first date for clarity
