### 2022-07-07 f18e639
docs: grayscale in parenthesis as it might also be RGB (https://github.com/sparks-baird/xtal2png/pull/111)
