### 2022-07-07 de15339
docs: correlaries does not exist in Merriam-Webster dictionary
