+++ reports/paper.md
@@ -87,7 +87,7 @@ take [Google's image-to-image diffusion model,
 Palette](https://iterative-refinement.github.io/palette/)
 [@sahariaPaletteImagetoImageDiffusion2022], which supports unconditional image
 generation, conditional inpainting, and conditional JPEG restoration tasks with
-correlaries for crystal generation, structure prediction, and structure relaxation,
+correlates for crystal generation, structure prediction, and structure relaxation,
 respectively. Rather than dig into the code and spending hours, days, or weeks modifying,
 debugging, and playing GitHub phone tag with the developers before you can (maybe) get
 preliminary results, `xtal2png` lets you get those results using the default
