### 2022-07-07 dc2b4e3
docs: add citation to SELFIES review/perspective
