### 2022-07-07 bf70b2f
touch-ups to kevin's changes
- drop grayscale
- material informatics --> materials informatics (https://en.wikipedia.org/wiki/Materials_informatics)
- mention the "multiple discovery" (really "multiple domain transfer") of Roost and CrabNet, i.e. sometimes the multiple year periods of domain transfer can result in significant overlapping work among multiple research groups
- correlates --> corresponding applications
