### 2022-07-07 62511f7
feat: implement click CLI
closes `xtal2png` without arguments throws `ValueError` #145
