+++ src/xtal2png/core.py
@@ -1,5 +1,4 @@
 """Crystal to PNG conversion core functions and scripts."""
-import argparse
 import logging
 import sys
 from functools import lru_cache
@@ -1188,74 +1187,6 @@ class XtalConverter:
         return data
 
 
-# ---- CLI ----
-# The functions defined in this section are wrappers around the main Python
-# API allowing them to be called directly from the terminal as a CLI
-# executable/script.
-
-
-def parse_args(args):
-    """Parse command line parameters.
-
-    Args:
-      args (List[str]): command line parameters as list of strings
-          (for example  ``["--help"]``).
-
-    Returns:
-      :obj:`argparse.Namespace`: command line parameters namespace
-    """
-    parser = argparse.ArgumentParser(description="Crystal to PNG encoder/decoder.")
-    parser.add_argument(
-        "--version",
-        action="version",
-        version="xtal2png {ver}".format(ver=__version__),
-    )
-    parser.add_argument(
-        "-p",
-        "--path",
-        dest="fpath",
-        help="Crystallographic information file (CIF) filepath (extension must be .cif or .CIF) or path to directory containing .cif files or processed PNG filepath or path to directory containing processed .png files (extension must be .png or .PNG). Assumes CIFs if --encode flag is used. Assumes PNGs if --decode flag is used.",  # noqa: E501
-        type=str,
-        metavar="STRING",
-    )
-    parser.add_argument(
-        "-s",
-        "--save-dir",
-        dest="save_dir",
-        default=".",
-        help="Directory to save processed PNG files or decoded CIFs to.",
-        type=str,
-        metavar="STRING",
-    )
-    parser.add_argument(
-        "--encode",
-        action="store_true",
-        help="Encode CIF files as PNG images.",
-    )
-    parser.add_argument(
-        "--decode",
-        action="store_true",
-        help="Decode PNG images as CIF files.",
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        dest="loglevel",
-        help="set loglevel to INFO",
-        action="store_const",
-        const=logging.INFO,
-    )
-    parser.add_argument(
-        "-vv",
-        "--very-verbose",
-        dest="loglevel",
-        help="set loglevel to DEBUG",
-        action="store_const",
-        const=logging.DEBUG,
-    )
-    return parser.parse_args(args)
-
-
 def setup_logging(loglevel):
     """Setup basic logging
 
@@ -1267,71 +1198,6 @@ def setup_logging(loglevel):
         level=loglevel, stream=sys.stdout, format=logformat, datefmt="%Y-%m-%d %H:%M:%S"
     )
 
-
-def main(args):
-    """Wrapper allowing :func:`XtalConverter()` :func:`xtal2png()` and
-    :func:`png2xtal()` methods to be called with string arguments in a CLI fashion.
-
-    Args:
-      args (List[str]): command line parameters as list of strings
-          (for example  ``["--verbose", "example.cif"]``).
-    """
-    args = parse_args(args)
-    setup_logging(args.loglevel)
-    _logger.debug("Beginning conversion to PNG format")
-
-    if args.encode and args.decode:
-        raise ValueError("Specify --encode or --decode, not both.")
-
-    if args.encode:
-        ext = ".cif"
-    elif args.decode:
-        ext = ".png"
-    else:
-        raise ValueError("Specify at least one of --encode or --decode")
-
-    if Path(args.fpath).suffix in [ext, ext.upper()]:
-        fpaths = [args.fpath]
-    elif path.isdir(args.fpath):
-        fpaths = glob(path.join(args.fpath, f"*{ext}"))
-        if fpaths == []:
-            raise ValueError(
[…]
