+++ setup.cfg
@@ -55,6 +55,7 @@ install_requires =
     plotly
     kaleido
     element-coder
+    click
 
 
 [options.packages.find]
@@ -81,7 +82,7 @@ testing =
 #     script_name = xtal2png.module:function
 # For example:
 console_scripts =
-    xtal2png = xtal2png.core:run
+    xtal2png = xtal2png.cli:cli
 # And any other entry points, for example:
 # pyscaffold.cli =
 #     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
