+++ tests/cli_test.py
@@ -1,30 +1,25 @@
 from os import path
 
-from xtal2png.core import main, parse_args
+from click.testing import CliRunner
 
-
-def test_parse_encode_args():
-    args = parse_args(["--encode", "--path", "src/xtal2png/utils/Zn2B2PbO6.cif"])
-    return args
-
-
-def test_parse_decode_args():
-    args = parse_args(
-        ["--decode", "--path", "data/external/preprocessed/Zn2B2PbO6.png"]
-    )
-    return args
+from xtal2png.cli import cli
 
 
 def test_encode_single():
     fpath = path.join("src", "xtal2png", "utils", "Zn2B2PbO6.cif")
     args = ["--encode", "--path", fpath, "--save-dir", "tmp"]
-    main(args)
+
+    runner = CliRunner()
+    result = runner.invoke(cli, args)
+    assert result.exit_code == 0
 
 
 def test_encode_dir():
     fpath = path.join("src", "xtal2png", "utils")
     args = ["--encode", "--path", fpath, "--save-dir", path.join("tmp", "tmp")]
-    main(args)
+    runner = CliRunner()
+    result = runner.invoke(cli, args)
+    assert result.exit_code == 0
 
 
 def test_decode_single():
@@ -32,21 +27,21 @@ def test_decode_single():
         "data", "preprocessed", "examples", "Zn8B8Pb4O24,volume=623,uid=b62a.png"
     )
     args = ["--decode", "--path", fpath, "--save-dir", "tmp"]
-    main(args)
+    runner = CliRunner()
+    result = runner.invoke(cli, args)
+    assert result.exit_code == 0
 
 
 def test_decode_dir():
     fpath = path.join("data", "preprocessed", "examples")
     args = ["--decode", "--path", fpath, "--save-dir", "tmp"]
-    main(args)
+    runner = CliRunner()
+    result = runner.invoke(cli, args)
+    assert result.exit_code == 0
 
 
 if __name__ == "__main__":
-    args = test_parse_encode_args()
-    args = test_parse_decode_args()
     test_encode_single()
     test_encode_dir()
     test_decode_single()
     test_decode_dir()
-
-    1 + 1
