### 2022-07-07 55f3a94
fix: also allow single file
