+++ docs/examples.md
@@ -4,10 +4,8 @@
 ```
 
 ```{toctree}
----
-maxdepth: 2
-caption: Contents:
----
+:maxdepth: 2
+
 notebooks/1.0-xtal2png-tutorial
 notebooks/2.0-materials-project-feature-ranges
 notebooks/2.0.1-materials-project-feature-ranges-conventional
