### 2022-07-08 fe955b2
docs: update readme cli example
