+++ docs/conf.py
@@ -77,6 +77,7 @@ extensions = [
     "sphinx_rtd_theme",
     "sphinx_copybutton",
     "nbsphinx",
+    "nbsphinx_link",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
