+++ docs/scripts/imagen_pytorch_example.md
@@ -1,0 +1,6 @@
+# Imagen PyTorch Example
+
+```{literalinclude} ../../scripts/imagen_pytorch_example.py
+:language: python
+:linenos:
+```
