+++ docs/scripts/denoising_diffusion_pytorch_example.md
@@ -1,0 +1,6 @@
+# Denoising Diffusion PyTorch Example
+
+```{literalinclude} ../../scripts/denoising_diffusion_pytorch_example.py
+:language: python
+:linenos:
+```
