### 2022-07-08 3656881
rename toctree entry from regression to classification
