+++ docs/examples.md
@@ -9,7 +9,7 @@
 notebooks/1.0-xtal2png-tutorial
 notebooks/2.0-materials-project-feature-ranges
 notebooks/2.0.1-materials-project-feature-ranges-conventional
-notebooks/2.1-xtal2png-cnn-regression
+notebooks/2.1-xtal2png-cnn-classification
 notebooks/2.2-xgboost-matbench-benchmark
 notebooks/3.0-denoising-diffusion
 notebooks/3.1-imagen-pytorch
