+++ docs/scripts/denoising_diffusion_pytorch_pretrained_sample.md
@@ -1,4 +1,4 @@
-# Denoising Diffusion PyTorch Pretrained Sample
+# Denoising Diffusion PyTorch Pretrained Sample (script)
 
 ```{literalinclude} ../../scripts/denoising_diffusion_pytorch_pretrained_sample.py
 :language: python
