### 2022-07-08 916bd6b
Create encoding_test.py
