### 2022-07-08 6a6eb86
add commented "compare with reference images" code
