### 2022-07-08 sparks-baird/xtal2png#146 @sgbaird
@PeterKraus I've [updated the index page with a blurb](https://xtal2png.readthedocs.io/en/latest/), [added references to the API in a Limitations and Design Considerations section](https://xtal2png.readthedocs.io/en/latest/readme.html#limitations-and-design-considerations) (as well as in the [basic usage section](https://xtal2png.readthedocs.io/en/latest/readme.html#example)), and split `xtal2png_test.py` into 4 modules each with a one-line docstring at the top:

![image](https://user-images.githubusercontent.com/45469701/177916854-fce26f71-956b-4f85-b5c9-4a5372c3b2ea.png)

I like your suggestion about comparing reference images as a category of testing and this is something I think I've considered before. I've added a test that saves the images to files, and then uses the loaded versions of those saved images for decoding.

As far as having fixed reference images to compare against, I'd like to hold off on this one until the default API becomes stable (i.e. `v1.0.0`), which I'm planning to do in conjunction with posting a preprint for a results manuscript. I'm planning to update many of the defaults based on results that I've yet to obtain and possibly along the way. I'll keep an issue open as a reminder of this. Does this seem reasonable?

Let me know if I missed something or if something seems off about the changes. Thanks for the feedback!
