### 2022-07-08 bec0c2d
store save_names as an attribute, and test loading from saved images
