### 2022-07-08 390f92e
fix typo (as --> and) and syntax error in 2 references for smiles/selfies
