+++ reports/paper.md
@@ -98,8 +98,8 @@ can (maybe) get preliminary results, `xtal2png` lets you get those results using
 default instructions on the repository, assuming the instructions can be run without
 error. While there are other invertible representations for crystal structures
 [@xieCrystalDiffusionVariational2022;@renInvertibleCrystallographicRepresentation2022a]
-as well as cross-domain conversions such as converting between molecules as strings
-[@weiningerSMILESChemicalLanguage1988,@selfies], to our knowledge, this is the first
+as well as cross-domain conversions such as converting between molecules and strings
+[@weiningerSMILESChemicalLanguage1988;@selfies], to our knowledge, this is the first
 package that enables conversion between a crystal structure and an image file format.
 
 ![(a) upscaled example image and (b) legend of the `xtal2png` encoding.\label{fig:example-and-legend}](figures/example-and-legend.png)
