### 2022-07-08 sparks-baird/xtal2png#144 @sgbaird
> * the reference to the software repo seems unnecessary here (and throughout the paper). The link to the repo is right there in the front-matter!

These have been removed.

> * the summary is not clear about what is being encoded into what. My understanding is that `xtal2png` can take a coordinate file and create a fingerprint / QR-code-like representation of that coordinate file. If that's the case, it should be more explicit in the summary.

> * streamlined results? Is the point here that it feeds well into image-based pipelines?

Clarified both the encoding and meaning of streamlined as follows:

If this is still off, lmk and I can take another stab.

> * Here, rather than to-from, I'd say between coordinate file and png representation (or something similar):

Used "between" instead of to/from

> * I am not sure I understand the relevance of Pallette. Please elaborate.

> * It would be helpful to discuss the technical limitations somewhere. Perhaps it's in the docs, but in my view it could be commented on in the paper:

Per your later comment in #146, I added a [section to the docs](https://xtal2png.readthedocs.io/en/latest/readme.html#limitations-and-design-considerations) and addressed the 3 points you brought up. Again, lmk if I'm missing something here.

> * I would rewrite this section, as a "potential" time saving is only potential.

Rewritten:

> * There is no mention of any other software that provides similar functionality. If there truly is not any encoder/decoder for crystal structure to image data, perhaps a short section discussing existing software for molecules and/or textual representation could be included.

I'm not aware of an encoder/decoder for crystal structure directly to an image file format. Added the following:
