### 2022-07-08 sparks-baird/xtal2png#146 @sgbaird
I also added an [examples section](https://xtal2png.readthedocs.io/en/latest/examples.html):

![image](https://user-images.githubusercontent.com/45469701/177920629-a6b703ac-f595-4d82-9301-dfb980b5c9df.png)

Let me know what you think!
