### 2022-07-08 4ce4fe5
fixup test, loading saved images
