### 2022-07-08 d3ca842
for consistency remove relative import. Ignore mypy error
mypy error: can't assign to method
