### 2022-07-08 35b8977
rm code graveyard in core.py
