### 2022-07-08 #4528 @sgbaird
@PeterKraus thanks for your feedback! I've addressed each of your comments for the [Software paper](https://github.com/sparks-baird/xtal2png/issues/144) and the [Documentation](https://github.com/sparks-baird/xtal2png/issues/146) issues. It's ready for a second look!

(p.s. I wasn't sure if I should be marking the tasks/checkboxes as done or not, so I left them unchecked)
