+++ CHANGELOG.md
@@ -1,5 +1,104 @@
 # CHANGELOG
 
+## Unreleased
+
+Changes unreleased.
+
+### Bug Fixes
+
+- general:
+  - fixup joss badge, remove leftover \autoref command ([0b4667a](https://github.com/sparks-baird/xtal2png/commit/0b4667a01da9944bc9786560a4098d8df5985c84)) ([#172](https://github.com/sparks-baird/xtal2png/pull/172))
+
+## [v0.8.0](https://github.com/sparks-baird/xtal2png/releases/tag/untagged-1cfe4c6287b8c846dc48) - 2022-07-08 07:13:50
+
+<!-- Release notes generated using configuration in .github/release.yml at v0.8.0 -->
+
+## What's Changed
+
+### Features
+* [WIP] feat: encode and decode using `element_coder` by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/117
+* replace volume with num_sites and use num_sites as a mask by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/141
+* allow for masking out redundant info from representation by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/138
+* public api / top-level imports since not using namespace packages by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/143
+
+### JOSS Paper
+* Add Kevin Jablonka @kjappelbaum and Colton Seegmiller @cseeg as co-authors to JOSS manuscript by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/148
+* Response to PeterKraus software paper review comments by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/147
+
+### Notebooks/Documentation
+* Update 2.1-xtal2png-cnn-classification.ipynb by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/158
+* jupyter notebooks and python scripts surfaced in docs via nbsphinx by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/151
+
+### Other Changes
+* png2xtal: if not isinstance(images, list) raise ValueError by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/137
+* max_sites error msg: note about encode/decode_cell_type kwargs by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/139
+* test_max_sites by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/140
+* mention XtalConverter as top-level API, and other links to API by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/150
+* move toctree to examples.md and use nbsphinx recommended format by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/152
+* docs: paper edits by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/153
+* fixup toctree syntax by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/155
+* nbsphinx_link and scripts in examples tab by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/156
+* add ipykernel as docs req by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/157
+* fixup files and content for nbsphinx_link by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/159
+* add title to imagen-pytorch nb by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/160
[…]
