### 2022-07-11 #4528 @dandavies99
Hi @sgbaird, this software is looking great. @PeterKraus - thanks for getting there first and carrying out such a thorough review! You've covered almost all of the main points I spotted already. I'll update comments in this thread and open specific issues as I go through the checklist. A few initial minor points from me (apologies if these are already being addressed in separate issues):

# Installation
- There are two conflicting sets of install instructions on the README: `conda create -n xtal2png -c conda-forge xtal2png m3gnet` at the top and `conda env create -n xtal2png -c conda-forge xtal2png` in the Installation section. The inclusion of `m3gnet` results in a _very_ long (possibly infinite at the time of writing 😉) environment solving time.  I understand m3gnet is optional, which is explained, but it might be worth not including this in the first installation instruction to save the user some frustration.
- The Example in the README doesn't work straight away for me. I think because `XtalConverter` is in `xtal2png.core` now, not in `xtal2png`. Similarly `example_structures` is now in `xtal2png.utils.data`.

# Docs
- Are fantastic. @PeterKraus has made some excellent suggestions.
- The "Getting Started, Output) section has the Original, Decoded and Relaxed Decoded outputs side by side regardless of page width. Could do with being one on top of the other.
