### 2022-07-11 #4528 @dandavies99
From my side, once the minor points I've mentioned above are addressed I will tick off Installation and Functionality. Other than that, I'm also happy and look forward to seeing this package in action - I'm sure it'll be very widely used.
