### 2022-07-15 sparks-baird/xtal2png#144 @PeterKraus
A couple more comments/suggestions:
- line 23: "in June 2017" would be more readable.
- line 30: I would write "for crystal structure applications, which was first introduced in a preprint (Kipf & Welling, 2016) and published in a peer-reviewed journal over a year later (Xie & Grossman 2017, 2018).
- line 33: I would spell out the need here: you want to help shorten this gap, with a focus on...
- line 38: instead of "tasks with corresponding applications", extend to "tasks, which are features with corresponding applications"
- line 42: "those results using the default instructions on the repository" perhaps "comparable results using the default parameters"
- line 61: merge paragraph with previous.
