### 2022-07-15 sparks-baird/xtal2png#179 @PeterKraus
Suggestion: Add CLI parameter for `max_sites`
When running `xtal2png` with a large crystal, this (expected) error appears:
```
(xtal2png) kraus@dorje:~/xtal2png::xtal2png --encode -p test-01.cif
/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/pymatgen/io/cif.py:1155: UserWarning: Issues encountered while parsing CIF: Some fractional coordinates rounded to ideal values to avoid issues with finite precision.
  warnings.warn("Issues encountered while parsing CIF: " + "\n".join(self.warnings))
  0%|                                                                                                                                                                    | 0/1 [00:00<?, ?it/s]
Traceback (most recent call last):
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/bin/xtal2png", line 10, in <module>
    sys.exit(run())
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 1141, in run
    main(sys.argv[1:])
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 1129, in main
    xc.xtal2png(fpaths, save=True)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 265, in xtal2png
    self.data, self.id_data, self.id_mapper = self.structures_to_arrays(S)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 573, in structures_to_arrays
    raise ValueError(
ValueError: crystal supplied with 240 sites, which is more than 52 sites. Remove crystal or increase `max_sites`.
```

It would be helpful to have a CLI option to increase `max_sites` with, rather than having to dig into the API.

Back-ping to https://github.com/openjournals/joss-reviews/issues/4528.
