### 2022-07-15 sparks-baird/xtal2png#180 @PeterKraus
Bug: `xtal2png` error with fractional occupancy
When `xtal2png` is executed with the attached file: [test-04.cif.txt](https://github.com/sparks-baird/xtal2png/files/9119471/test-04.cif.txt) (.txt added so GH does not complain), the following error appears:

```
(xtal2png) kraus@dorje:~/xtal2png::xtal2png --encode -p test-04.cif -vv
[2022-07-15 10:59:42] DEBUG:xtal2png.core:Beginning conversion to PNG format
  0%|                                                                                                                                                                    | 0/1 [00:00<?, ?it/s]
Traceback (most recent call last):
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/pymatgen/core/structure.py", line 267, in atomic_numbers
    return tuple(site.specie.Z for site in self)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/pymatgen/core/structure.py", line 267, in <genexpr>
    return tuple(site.specie.Z for site in self)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/pymatgen/core/sites.py", line 79, in __getattr__
    raise AttributeError(a)
AttributeError: specie. Did you mean: 'species'?

Traceback (most recent call last):
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/bin/xtal2png", line 10, in <module>
    sys.exit(run())
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 1141, in run
    main(sys.argv[1:])
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 1129, in main
    xc.xtal2png(fpaths, save=True)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 265, in xtal2png
    self.data, self.id_data, self.id_mapper = self.structures_to_arrays(S)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/xtal2png/core.py", line 571, in structures_to_arrays
    n_sites = len(s.atomic_numbers)
  File "/home/kraus/Applications/miniconda3/envs/xtal2png/lib/python3.10/site-packages/pymatgen/core/structure.py", line 269, in atomic_numbers
    raise AttributeError("atomic_numbers available only for ordered Structures")
AttributeError: atomic_numbers available only for ordered Structures
```

This is not a very useful error message. I believe it's caused by the fractional occupancy in the crystal. If it is so, such cases should be handled more gracefully.

Back-ping to https://github.com/openjournals/joss-reviews/issues/4528
