### 2022-07-15 sparks-baird/xtal2png#180 @kjappelbaum
Thanks for the bug report!
