### 2022-07-15 #4528 @PeterKraus
Alright, I have made two more issues, but they are not a blocker for me accepting. After you address the comments in https://github.com/sparks-baird/xtal2png/issues/144, feel free to close the issue. The documentation issue, https://github.com/sparks-baird/xtal2png/issues/146, can be closed already.

@rkurchin, do I need to do anything with the bot to recommend accept?
