### 2022-07-17 27c39ec
feat: add `max_sites` to CLI
Fixes #179
