### 2022-07-17 6ed5805
chore: added a large MOF as test case
