### 2022-07-17 449f949
fix: catch partial occupancy
Fixes #180
