### 2022-07-17 9e6bcad
fix: add fixture decorator
