### 2022-07-17 6af92ba
fix: add test file of disordered structure
