### 2022-07-17 70a1c7e
chore: also add CLI test
