+++ tests/cli_test.py
@@ -4,6 +4,8 @@ from click.testing import CliRunner
 
 from xtal2png.cli import cli
 
+THIS_DIR = path.dirname(path.abspath(__file__))
+
 
 def test_encode_single():
     fpath = path.abspath(path.join("src", "xtal2png", "utils", "Zn2B2PbO6.cif"))
@@ -39,6 +41,16 @@ def test_decode_dir():
     runner = CliRunner()
     result = runner.invoke(cli, args)
     assert result.exit_code == 0
+
+
+def test_disordered_structure():
+    fpath = path.join(THIS_DIR, "test_files" "disordered_structure.cif")
+    args = ["--encode", "--path", fpath, "--save-dir", "tmp"]
+    runner = CliRunner()
+    result = runner.invoke(cli, args, catch_exceptions=True)
+    assert result.exit_code == 1
+    assert isinstance(result.exception, ValueError)
+    assert "xtal2png does not support disordered structures." in str(result.exception)
 
 
 if __name__ == "__main__":
