### 2022-07-17 f02d4b4
fix: broken instance check
