### 2022-07-17 dfb1f36
docs: Jun -> June
