### 2022-07-17 16ae345
paper: not all italics for chemical formula
