### 2022-07-25 sparks-baird/xtal2png#184 @kjappelbaum
JOSS paper review - Installation docs 
> There are two conflicting sets of install instructions on the README: conda create -n xtal2png -c conda-forge xtal2png m3gnet at the top and conda env create -n xtal2png -c conda-forge xtal2png in the Installation section. The inclusion of m3gnet results in a very long (possibly infinite at the time of writing 😉) environment solving time. I understand m3gnet is optional, which is explained, but it might be worth not including this in the first installation instruction to save the user some frustration.

- [x] Example
> The Example in the README doesn't work straight away for me. I think because XtalConverter is in xtal2png.core now, not in xtal2png. Similarly example_structures is now in xtal2png.utils.data.

from https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1180113793
