### 2022-07-25 sparks-baird/xtal2png#185 @kjappelbaum
JOSS paper review - Docs
> On ReadTheDocs, The "Getting Started, Output" section has the Original, Decoded and Relaxed Decoded outputs side by side regardless of page width. Could do with being one on top of the other.

from https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1180113793
