### 2022-07-25 0641b7c
docs: comment out pending conda installation
