### 2022-07-25 sparks-baird/xtal2png#184 @kjappelbaum
Running the example actually produces the following

```python
In [9]: for i, s in enumerate(example_structures):
   ...:     print(s.get_space_group_info())
   ...:
('Pccn', 56)
Input In [9], in <cell line: 1>()
      1 for i, s in enumerate(example_structures):
----> 2     print(s.get_space_group_info())

File ~/miniconda3/envs/xtal2png/lib/python3.8/site-packages/pymatgen/core/structure.py:1013, in IStructure.get_space_group_info(self, symprec, angle_tolerance)
   1010 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
   1012 a = SpacegroupAnalyzer(self, symprec=symprec, angle_tolerance=angle_tolerance)
-> 1013 return a.get_space_group_symbol(), a.get_space_group_number()

File ~/miniconda3/envs/xtal2png/lib/python3.8/site-packages/pymatgen/symmetry/analyzer.py:97, in SpacegroupAnalyzer.get_space_group_symbol(self)
     90 def get_space_group_symbol(self):
     91     """
     92     Get the spacegroup symbol (e.g., Pnma) for structure.
     93
     94     Returns:
     95         (str): Spacegroup symbol for structure.
     96     """
---> 97     return self._space_group_data["international"]

TypeError: 'NoneType' object is not subscriptable
```
