### 2022-07-25 sparks-baird/xtal2png#184 @kjappelbaum
apparently, because `spglib.get_symmetry_dataset` returns `None` for this structure.
