### 2022-07-25 dbb2e68
fix: if space group cannot be found, return 0
