### 2022-07-25 cd8db1a
JOSS paper review - Docs Fixes #185
