### 2022-07-25 sparks-baird/xtal2png#184 @dandavies99
Hi @kjappelbaum, @sgbaird - I'm still getting import errors trying the example as in the README. I think it should be:

```diff
- from xtal2png import XtalConverter, example_structures
+ from xtal2png.core import XtalConverter
+ from xtal2png.utils.data import example_structures
```
