### 2022-07-25 #4528 @dandavies99
Looking good. Just one query which I've commented on [this issue](https://github.com/sparks-baird/xtal2png/issues/184).
