### 2022-07-25 sparks-baird/xtal2png#184 @kjappelbaum
thanks!

did you install from the `HEAD` or did you test some specific release?  In principle, those things should be exported here ... https://github.com/sparks-baird/xtal2png/blob/464e767207ee281fc19cfaa12d87f17ff6b988c4/src/xtal2png/__init__.py#L20-L21

Anyhow, let's add a test for this example.
