### 2022-07-25 sparks-baird/xtal2png#184 @sgbaird
Strange, I would have thought it would work based on https://github.com/sparks-baird/xtal2png/blob/main/src/xtal2png/__init__.py#L20-L21. Which version are you using and how are you doing the installation?
