### 2022-07-25 sparks-baird/xtal2png#184 @sgbaird
@dandavies99 ah, right. This should be corrected in `v0.8.0`, but conda-forge is still at `v0.7.0` while @kjappelbaum is working on getting `element-coder` up on conda-forge. I'll change the directions back to what you suggested in the meantime, and then change it back once `conda-forge` is up to `v0.8.0`.
