### 2022-07-25 sparks-baird/xtal2png#184 @dandavies99
Great that makes sense! It's a v minor point, but just thinking about a novice user playing with the code for the first time; they'll probably install using conda then try and run the first example, so good to have those things consistent 😄
