+++ README.md
@@ -40,7 +40,8 @@ class with and without optional relaxation via
 
 ```python
 # example_structures is a list of `pymatgen.core.structure.Structure` objects
-from xtal2png import XtalConverter, example_structures
+from xtal2png.core import XtalConverter
+from xtal2png.utils.data example_structures
 
 xc = XtalConverter(relax_on_decode=True)
 data = xc.xtal2png(example_structures, show=True, save=True)
