### 2022-07-25 89fac68
chore: add doctest
