+++ README.md
@@ -40,16 +40,18 @@ class with and without optional relaxation via
 
 ```python
 # example_structures is a list of `pymatgen.core.structure.Structure` objects
-from xtal2png import XtalConverter, example_structures
+>>> from xtal2png import XtalConverter, example_structures
 
-xc = XtalConverter(relax_on_decode=True)
-data = xc.xtal2png(example_structures, show=True, save=True)
-relaxed_decoded_structures = xc.png2xtal(data, save=False)
+>>> xc = XtalConverter(relax_on_decode=True)
+>>> data = xc.xtal2png(example_structures, show=True, save=True)
+>>> relaxed_decoded_structures = xc.png2xtal(data, save=False)
 
 
-xc = XtalConverter(relax_on_decode=False)  # default behavior
-data = xc.xtal2png(example_structures, show=True, save=True)
-decoded_structures = xc.png2xtal(data, save=False)
+>>> xc = XtalConverter(relax_on_decode=False)  # default behavior
+>>> data = xc.xtal2png(example_structures, show=True, save=True)
+>>> decoded_structures = xc.png2xtal(data, save=False)
+
+assert len(decoded_structures) == len(example_structures)
 ```
 
 ### Output
