### 2022-07-25 6fe3aac
chore: add doctest
