+++ tox.ini
@@ -64,6 +64,7 @@ deps =
     # ^  requirements.txt shared with Read The Docs
 commands =
     sphinx-build --color -b {env:BUILD} -d "{env:BUILDDIR}/doctrees" "{env:DOCSDIR}" "{env:BUILDDIR}/{env:BUILD}" {posargs}
+    doctest: python -m doctest README.md
 
 
 [testenv:publish]
