### 2022-07-28 #4528 @rkurchin
Conceptual comment:
In your Statement of Need, you compare crystal graph convolution (e.g. `cgcnn.py` to what you've done here, but I would argue there's a meaningful conceptual distinction there, in that CGCNN's are their own model architecture taking in a crystal graph as input. What you've done is create a new kind of data representation as images (which one could think of as graphs with a very regular local adjacency structure), which allows invocation of a whole bunch of image-based architectures _without_ needing to generalize them to arbitrary graphs, the way GCNN's are a generalization of CNN's. I would consider whether another example/framing might help clarify this...
(Curious also for the reviewers' takes on this critique if they care to chime in, @sgbaird and @dandavies99)

Some editorial suggestions:
* line 10: "natural language" -> "natural language processing"
* 11: capitalize and add a comma after Transformers (also capitalize in other places it shows up, e.g. line 23)
* 15: delete "respectively"
* 22: be consistent with citation formatting, put Vaswani et al. inside citation
* 23: semicolon should be a comma (second clause could not be a sentence on its own)
* lines 31 and 33: something weird is happening with the double-parentheses for citation
* 60: either add a comma after before "which" or change it to "that"

P.S. I think this is a pretty nifty idea and may be trying it out in some of my own work!
