### 2022-07-28 #4528 @sgbaird
> Conceptual comment:
> In your Statement of Need, you compare crystal graph convolution (e.g. `cgcnn.py` to what you've done here, but I would argue there's a meaningful conceptual distinction there, in that CGCNN's are their own model architecture taking in a crystal graph as input. What you've done is create a new kind of data representation as images (which one could think of as graphs with a very regular local adjacency structure), which allows invocation of a whole bunch of image-based architectures _without_ needing to generalize them to arbitrary graphs, the way GCNN's are a generalization of CNN's. I would consider whether another example/framing might help clarify this...
> (Curious also for the reviewers' takes on this critique if they care to chime in, @sgbaird and @dandavies99)

That's a good point. CGCNN was a generalization of a CNN and affected how the data was represented. This is more about screening models using a new data representation. E.g. try `xtal2png` with `imagen-pytorch`, a GAN, and a VAE, and see which one does better, then decide whether to spend more time making a generalized data representation + model combination for the one(s) that perform best. Hence the screening.

Maybe I could use https://github.com/aspuru-guzik-group/chemical_vae as an example, which I think was the first implementation of VAEs to molecules. At least it was probably the first widely publicized/popular one. VAEs introduced in 2013, Chemical VAE implemented in ~2018.

I could also use [CDVAE](https://github.com/txie-93/cdvae) as an example which was first online (arXiv) Oct 2021, introduced in 2015.

Your coment about it being "a graph with a very regular local adjacency structure" reminded me of [this twitter comment](https://twitter.com/rekumar_/status/1538404893731409920):

> The interesting part to me is that pixels carry different info (features vs positions). Image convolutions/kernels are funky in this scenario. Feels like a messier graph CNN where edges and node features are mixed up… curious how it goes!

> Some editorial suggestions:
>
> * line 10: "natural language" -> "natural language processing"
> * 11: capitalize and add a comma after Transformers (also capitalize in other places it shows up, e.g. line 23)
> * 15: delete "respectively"
> * 22: be consistent with citation formatting, put Vaswani et al. inside citation
> * 23: semicolon should be a comma (second clause could not be a sentence on its own)
> * lines 31 and 33: something weird is happening with the double-parentheses for citation
> * 60: either add a comma after before "which" or change it to "that"

Great, will get these implemented.

> P.S. I think this is a pretty nifty idea and may be trying it out in some of my own work!

Glad to hear!
