### 2022-07-28 4b25c6a
add bibtex entries for VAEs and DDPMs
https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1198640405
