### 2022-07-28 2c2026b
add bibtex entries for VAEs and DDPMs
https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1198640405
