+++ reports/paper.bib
@@ -1,236 +1,322 @@
 
+@article{gomez-bombarelliAutomaticChemicalDesign2018,
+  title = {Automatic Chemical Design Using a Data-Driven Continuous Representation of Molecules},
+  author = {{G{\'o}mez-Bombarelli}, Rafael and Wei, Jennifer N. and Duvenaud, David and {Hern{\'a}ndez-Lobato}, Jos{\'e} Miguel and {S{\'a}nchez-Lengeling}, Benjam{\'i}n and Sheberla, Dennis and {Aguilera-Iparraguirre}, Jorge and Hirzel, Timothy D. and Adams, Ryan P. and {Aspuru-Guzik}, Al{\'a}n},
+  year = {2018},
+  month = feb,
+  journal = {ACS Cent. Sci.},
+  volume = {4},
+  number = {2},
+  eprint = {1610.02415},
+  eprinttype = {arxiv},
+  primaryclass = {physics},
+  pages = {268--276},
+  issn = {2374-7943, 2374-7951},
+  doi = {10.1021/acscentsci.7b00572},
+  abstract = {We report a method to convert discrete representations of molecules to and from a multidimensional continuous representation. This generative model allows efficient search and optimization through open-ended spaces of chemical compounds.},
+  archiveprefix = {arXiv},
+  langid = {english},
+  keywords = {Computer Science - Machine Learning,Physics - Chemical Physics},
+  file = {C\:\\Users\\sterg\\Zotero\\storage\\2JMZE9VY\\Gómez-Bombarelli et al. - 2018 - Automatic chemical design using a data-driven cont.pdf}
+}
+
 @article{goodallPredictingMaterialsProperties2019,
-  title     = {Predicting Materials Properties without Crystal Structure: {{Deep}} Representation Learning from Stoichiometry},
-  author    = {Goodall, Rhys E. A. and Lee, Alpha A.},
-  year      = {2019},
+  title = {Predicting Materials Properties without Crystal Structure: {{Deep}} Representation Learning from Stoichiometry},
+  author = {Goodall, Rhys E. A. and Lee, Alpha A.},
+  year = {2019},
   publisher = {{arXiv}},
-  doi       = {10.48550/ARXIV.1910.00617},
+  doi = {10.48550/ARXIV.1910.00617},
   copyright = {arXiv.org perpetual, non-exclusive license},
-  keywords  = {Computational Physics (physics.comp-ph),FOS: Computer and information sciences,FOS: Physical sciences,Machine Learning (cs.LG),Materials Science (cond-mat.mtrl-sci)}
+  keywords = {Computational Physics (physics.comp-ph),FOS: Computer and information sciences,FOS: Physical sciences,Machine Learning (cs.LG),Materials Science (cond-mat.mtrl-sci)}
 }
 
 @article{goodallPredictingMaterialsProperties2020,
-  title         = {Predicting Materials Properties without Crystal Structure: {{Deep}} Representation Learning from Stoichiometry},
-  shorttitle    = {Predicting Materials Properties without Crystal Structure},
-  author        = {Goodall, Rhys E. A. and Lee, Alpha A.},
-  year          = {2020},
-  month         = dec,
-  journal       = {Nature Communications},
-  volume        = {11},
-  number        = {1},
-  eprint        = {1910.00617},
-  eprinttype    = {arxiv},
-  primaryclass  = {cond-mat, physics:physics},
-  pages         = {6280},
-  issn          = {2041-1723},
-  doi           = {10.1038/s41467-020-19964-7},
[…]
