### 2022-07-28 8c13d7c
mention VAEs and DDPMs
Addressing https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1198585724
