+++ reports/paper.md
@@ -71,8 +71,8 @@ image-based machine learning models applied to crystal structures.
 
 Using a state-of-the-art method in a separate domain with a custom data representation
 is often an expensive and drawn-out process. For example, @vaswaniAttentionAllYou2017
-introduced the revolutionary natural language processing transformer architecture in June
-2017; yet the application of transformers to the adjacent domain of materials
+introduced the revolutionary natural language processing transformer architecture in
+June 2017; yet the application of transformers to the adjacent domain of materials
 informatics (chemical-formula-based predictions) was not publicly realized until late
 2019 [@goodallPredictingMaterialsProperties2019], approximately two-and-a-half years
 later, with peer-reviewed publications dating to late 2020
@@ -82,11 +82,17 @@ slightly later public release [@wangCompositionallyrestrictedAttentionbasedNetwo
 and publication [@wangCompositionallyRestrictedAttentionbased2021] dates. Another
 example of a state-of-the-art algorithm domain transfer is refactoring image-processing
 models for crystal structure applications, which was first introduced in a preprint
-([@kipfSemisupervisedClassificationGraph2016]) and published with application for materials' property prediction
-in a peer-reviewed journal over a year later ([@xieCrystalGraphConvolutional2018]).
-Here, we focus on the latter application: state-of-the-art domain transfer
-from image processing to crystal structure to enable materials science practitioners
-to leverage the most advanced image processing models for materials' property prediction and inverse design.
+([@kipfSemisupervisedClassificationGraph2016]) and published with application for
+materials' property prediction in a peer-reviewed journal over a year later
+([@xieCrystalGraphConvolutional2018]). Similarly, variational autoencoders (VAEs) were
+introduced in 2013 ([@kingmaAutoEncodingVariationalBayes2014a]) and implemented for molecules in 2016
+([@gomez-bombarelliAutomaticChemicalDesign2018]) and denoising diffusion probabilistic models (DDPMs) were
+introduced in 2015 ([@sohl-dicksteinDeepUnsupervisedLearning2015]) and implemented for
+crystal structures in 2021 ([@xieCrystalDiffusionVariational2021]). Here, we focus on
+the latter application: state-of-the-art domain transfer (especially of generative
+models) from image processing to crystal structure to enable materials science
+practitioners to leverage the most advanced image processing models for materials'
+property prediction and inverse design.
 
 `xtal2png` is a Python package that allows you to convert between a crystal structure
 and a PNG image for direct use with image-based machine learning models. Let's take
