### 2022-07-28 87eb902
address editorial suggestions
https://github.com/openjournals/joss-reviews/issues/4528#issuecomment-1198585724

line 10: "natural language" -> "natural language processing"
11: capitalize and add a comma after Transformers (also capitalize in other places it shows up, e.g. line 23)
15: delete "respectively"
22: be consistent with citation formatting, put Vaswani et al. inside citation
23: semicolon should be a comma (second clause could not be a sentence on its own)
lines 31 and 33: something weird is happening with the double-parentheses for citation
60: either add a comma after before "which
[…]
