+++ reports/paper.md
@@ -54,12 +54,12 @@ bibliography: paper.bib
 # Summary
 
 The latest advances in machine learning are often in natural language processing such as with long
-short-term memory networks (LSTMs) and transformers or image processing such as with
+short-term memory networks (LSTMs) and Transformers, or image processing such as with
 generative adversarial networks (GANs), variational autoencoders (VAEs), and guided
 diffusion models. `xtal2png` encodes and decodes crystal structures via PNG
 images (see e.g. \autoref{fig:64-bit}) by writing and reading the necessary information
 for crystal reconstruction (unit cell, atomic elements, atomic coordinates) as a square
-matrix of numbers, respectively. This is akin to making/reading a QR code for crystal
+matrix of numbers. This is akin to making/reading a QR code for crystal
 structures, where the `xtal2png` representation is an invertible representation. The
 ability to feed these images directly into image-based pipelines allows you, as a
 materials informatics practitioner, to get streamlined results for new state-of-the-art
@@ -70,9 +70,9 @@ image-based machine learning models applied to crystal structures.
 # Statement of need
 
 Using a state-of-the-art method in a separate domain with a custom data representation
-is often an expensive and drawn-out process. For example, @vaswaniAttentionAllYou2017
-introduced the revolutionary natural language processing transformer architecture in
-June 2017; yet the application of transformers to the adjacent domain of materials
+is often an expensive and drawn-out process. For example, [@vaswaniAttentionAllYou2017]
+introduced the revolutionary natural language processing Transformer architecture in
+June 2017, yet the application of Transformers to the adjacent domain of materials
 informatics (chemical-formula-based predictions) was not publicly realized until late
 2019 [@goodallPredictingMaterialsProperties2019], approximately two-and-a-half years
 later, with peer-reviewed publications dating to late 2020
@@ -82,14 +82,14 @@ slightly later public release [@wangCompositionallyrestrictedAttentionbasedNetwo
 and publication [@wangCompositionallyRestrictedAttentionbased2021] dates. Another
 example of a state-of-the-art algorithm domain transfer is refactoring image-processing
 models for crystal structure applications, which was first introduced in a preprint
-([@kipfSemisupervisedClassificationGraph2016]) and published with application for
+[@kipfSemisupervisedClassificationGraph2016] and published with application for
 materials' property prediction in a peer-reviewed journal over a year later
-([@xieCrystalGraphConvolutional2018]). Similarly, variational autoencoders (VAEs) were
-introduced in 2013 ([@kingmaAutoEncodingVariationalBayes2014a]) and implemented for
-molecules in 2016 ([@gomez-bombarelliAutomaticChemicalDesign2018]) and denoising
+[@xieCrystalGraphConvolutional2018]. Similarly, variational autoencoders (VAEs) were
+introduced in 2013 [@kingmaAutoEncodingVariationalBayes2014a] and implemented for
+molecules in 2016 [@gomez-bombarelliAutomaticChemicalDesign2018] and denoising
 diffusion probabilistic models (DDPMs) were introduced in 2015
-([@sohl-dicksteinDeepUnsupervisedLearning2015]) and implemented for crystal structures
-in 2021 ([@xieCrystalDiffusionVariational2021]). Here, we focus on the latter
+[@sohl-dicksteinDeepUnsupervisedLearning2015] and implemented for crystal structures
+in 2021 [@xieCrystalDiffusionVariational2021]. Here, we focus on the latter
 application: state-of-the-art domain transfer (especially of generative models) from
 image processing to crystal structure to enable materials science practitioners to
 leverage the most advanced image processing models for materials' property prediction
@@ -128,7 +128,7 @@ numerical values as PNG images (allowable values are integers between 0 and
[…]
