### 2022-07-28 a8e7550
remove redundant abbreviation definition
