+++ reports/paper.bib
@@ -1,10 +1,20 @@
 
+@article{gomez-bombarelliAutomaticChemicalDesign2016,
+  title = {Automatic Chemical Design Using a Data-Driven Continuous Representation of Molecules},
+  author = {{G{\'o}mez-Bombarelli}, Rafael and Wei, Jennifer N. and Duvenaud, David and {Hern{\'a}ndez-Lobato}, Jos{\'e} Miguel and {S{\'a}nchez-Lengeling}, Benjam{\'i}n and Sheberla, Dennis and {Aguilera-Iparraguirre}, Jorge and Hirzel, Timothy D. and Adams, Ryan P. and {Aspuru-Guzik}, Al{\'a}n},
+  year = {2016},
+  publisher = {{arXiv}},
+  doi = {10.48550/ARXIV.1610.02415v1},
+  copyright = {arXiv.org perpetual, non-exclusive license},
+  keywords = {Chemical Physics (physics.chem-ph),FOS: Computer and information sciences,FOS: Physical sciences,Machine Learning (cs.LG)}
+}
+
 @article{gomez-bombarelliAutomaticChemicalDesign2018,
   title = {Automatic Chemical Design Using a Data-Driven Continuous Representation of Molecules},
   author = {{G{\'o}mez-Bombarelli}, Rafael and Wei, Jennifer N. and Duvenaud, David and {Hern{\'a}ndez-Lobato}, Jos{\'e} Miguel and {S{\'a}nchez-Lengeling}, Benjam{\'i}n and Sheberla, Dennis and {Aguilera-Iparraguirre}, Jorge and Hirzel, Timothy D. and Adams, Ryan P. and {Aspuru-Guzik}, Al{\'a}n},
   year = {2018},
   month = feb,
-  journal = {ACS Cent. Sci.},
+  journal = {ACS Central Science},
   volume = {4},
   number = {2},
   eprint = {1610.02415},
@@ -36,7 +46,7 @@
   author = {Goodall, Rhys E. A. and Lee, Alpha A.},
   year = {2020},
   month = dec,
-  journal = {Nat Commun},
+  journal = {Nature Communications},
   volume = {11},
   number = {1},
   eprint = {1910.00617},
@@ -216,7 +226,7 @@
   author = {Wang, Anthony Yu-Tung and Kauwe, Steven K. and Murdock, Ryan J. and Sparks, Taylor D.},
   year = {2021},
   month = dec,
-  journal = {npj Comput Mater},
+  journal = {npj Computational Materials},
   volume = {7},
   number = {1},
   pages = {77},
@@ -241,7 +251,7 @@
   author = {Weininger, David},
   year = {1988},
   month = feb,
-  journal = {J. Chem. Inf. Comput. Sci.},
+  journal = {Journal of Chemical Information and Computer Sciences},
   volume = {28},
   number = {1},
   pages = {31--36},
@@ -299,7 +309,7 @@
   author = {Xie, Tian and Grossman, Jeffrey C.},
   year = {2018},
   month = apr,
-  journal = {Phys. Rev. Lett.},
+  journal = {Physical Review Letters},
   volume = {120},
   number = {14},
   eprint = {1710.10324},
