+++ reports/paper.md
@@ -84,13 +84,12 @@ example of a state-of-the-art algorithm domain transfer is refactoring image-pro
 models for crystal structure applications, which was first introduced in a preprint
 [@kipfSemisupervisedClassificationGraph2016] and published with application for
 materials' property prediction in a peer-reviewed journal over a year later
-[@xieCrystalGraphConvolutional2018]. Similarly, VAEs were
-introduced in 2013 [@kingmaAutoEncodingVariationalBayes2014a] and implemented for
-molecules in 2016 [@gomez-bombarelliAutomaticChemicalDesign2018] and denoising
-diffusion probabilistic models (DDPMs) were introduced in 2015
-[@sohl-dicksteinDeepUnsupervisedLearning2015] and implemented for crystal structures
-in 2021 [@xieCrystalDiffusionVariational2021]. Here, we focus on the latter
-application: state-of-the-art domain transfer (especially of generative models) from
+[@xieCrystalGraphConvolutional2018]. Similarly, VAEs were introduced in 2013
+[@kingmaAutoEncodingVariationalBayes2014a] and implemented for molecules in 2016
+[@gomez-bombarelliAutomaticChemicalDesign2016], and denoising diffusion probabilistic
+models (DDPMs) were introduced in 2015 [@sohl-dicksteinDeepUnsupervisedLearning2015] and
+implemented for crystal structures in 2021 [@xieCrystalDiffusionVariational2021]. Here,
+we focus on state-of-the-art domain transfer (especially of generative models) from
 image processing to crystal structure to enable materials science practitioners to
 leverage the most advanced image processing models for materials' property prediction
 and inverse design.
