### 2022-07-28 e106b06
docs(CHANGELOG): update release notes
