+++ CHANGELOG.md
@@ -1,5 +1,57 @@
 # CHANGELOG
 
+## [v0.9.1](https://github.com/sparks-baird/xtal2png/releases/tag/v0.9.1) - 2022-07-28 22:06:22
+
+<!-- Release notes generated using configuration in .github/release.yml at v0.9.1 -->
+
+## What's Changed
+* cnn-classification notebook (Created using Colaboratory) by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/175
+* add regression notebook and change classification Colab badge to markdown for compatibility with nbsphinx by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/178
+* fix: catch partial occupancy by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/182
+* fix: spacegroup fallback and installation docs  by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/186
+* feat: add `max_sites` to CLI by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/181
+
+### Other Changes
+* fixup joss badge, remove leftover \autoref command by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/172
+* split large block of README text into subsections by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/171
+* convert colab badges from html to markdown for nbsphinx compatibility by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/173
+* Revert "cnn-classification notebook (Created using Colaboratory)" by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/176
+* Created using Colaboratory by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/177
+* paper: update software paper by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/183
+* JOSS paper review - Docs by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/187
+* add Berend as co-author by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/191
+* Revert imports until conda-forge is bumped by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/189
+* response to rkurchin's editorial review by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/192
+
+
+**Full Changelog**: https://github.com/sparks-baird/xtal2png/compare/v0.8.0...v0.9.1
+
+### Feature
+
+- general:
+  - add `max_sites` to CLI ([27c39ec](https://github.com/sparks-baird/xtal2png/commit/27c39ec7310525cfe8205cadcbdd2add7e8e964f)) ([#181](https://github.com/sparks-baird/xtal2png/pull/181))
+
+### Bug Fixes
+
+- general:
+  - if space group cannot be found, return 0 ([dbb2e68](https://github.com/sparks-baird/xtal2png/commit/dbb2e68b7a380d9ef03b2ce3b1fb9452d3a14f90)) ([#186](https://github.com/sparks-baird/xtal2png/pull/186))
+  - wrong filepath in text ([5e151ac](https://github.com/sparks-baird/xtal2png/commit/5e151ac03767ab157eb927ff767957bd9d28aae1)) ([#182](https://github.com/sparks-baird/xtal2png/pull/182))
+  - broken instance check ([f02d4b4](https://github.com/sparks-baird/xtal2png/commit/f02d4b433285a7f503770f688a5aa5bef47dac55)) ([#182](https://github.com/sparks-baird/xtal2png/pull/182))
[…]
