### 2022-07-28 sparks-baird/xtal2png#193 @sgbaird
Any acknowledgements that need to be added to `paper.md`?
@kjappelbaum (e.g. grant award number)
