### 2022-07-28 sparks-baird/xtal2png#193 @sgbaird
@cseeg @hasan-sayeed @michaeldalverson

https://github.com/sparks-baird/xtal2png/blob/main/reports/paper.md#acknowledgements
