### 2022-07-28 b9985cf
update authors, including adding Faris Khan
@faris-k
