### 2022-07-28 sparks-baird/xtal2png#193 @sp8rks
Those acknowledgements are fine by me.

On Thu, Jul 28, 2022, 6:35 PM Sterling Baird ***@***.***>
wrote:

> @cseeg <https://github.com/cseeg> @hasan-sayeed
> <https://github.com/hasan-sayeed> @michaeldalverson
> <https://github.com/michaeldalverson> @sp8rks <https://github.com/sp8rks>
> @faris-k <https://github.com/faris-k>
>
>
> https://github.com/sparks-baird/xtal2png/blob/main/reports/paper.md#acknowledgements
>
