### 2022-07-28 #4528 @sgbaird
@rkurchin thanks for the great feedback. I addressed your comments. Lmk if this looks OK.

Looking back over the manuscript, I realized I probably need to add to the acknowledgments section (https://github.com/sparks-baird/xtal2png/issues/193) and will post the final version and DOI after that.
