### 2022-07-28 sparks-baird/xtal2png#193 @sgbaird
@sp8rks do we acknowledge ReUSE? Also, not sure if the topological insulators grant should be acknowledged for Hasan.
