### 2022-07-28 sparks-baird/xtal2png#193 @sp8rks
You are right! Please acknowledge both.

On Thu, Jul 28, 2022, 7:33 PM Sterling Baird ***@***.***>
wrote:

> @sp8rks <https://github.com/sp8rks> do we acknowledge ReUSE? Also, not
> sure if the topological insulators grant should be acknowledged for Hasan.
>
