### 2022-07-29 sparks-baird/xtal2png#193 @sgbaird
@cseeg and @hasan-sayeed, will you each make a PR adding the appropriate statement to the acknowledgments section?
