### 2022-07-29 3d4acfb
try triggering doctests
