+++ .github/workflows/ci.yml
@@ -83,6 +83,8 @@ jobs:
           github-token: ${{ secrets.github_token }}
           flag-name: ${{ matrix.platform }} - py${{ matrix.python }}
           parallel: true
+      - name: Run doctest
+        run: pipx run tox -e doctest
 
   finalize:
     needs: test
