+++ .github/workflows/ci.yml
@@ -84,7 +84,7 @@ jobs:
           flag-name: ${{ matrix.platform }} - py${{ matrix.python }}
           parallel: true
       - name: Run doctest
-        run: pipx run tox -e doctest
+        run: pipx run tox -e doctests
 
   finalize:
     needs: test
