### 2022-07-29 1f91689
run doctest direclty
