### 2022-07-29 0c78f20
use pytest to run doctests is easiest
