+++ setup.cfg
@@ -96,11 +96,14 @@ console_scripts =
 addopts =
     --cov xtal2png --cov-report term-missing
     --verbose
+    --doctest-modules
+    --doctest-glob "README.md"
+    --doctest-continue-on-failure
 norecursedirs =
     dist
     build
     .tox
-testpaths = tests
+testpaths = tests README.md
 # Use pytest markers to select/deselect specific tests
 # markers =
 #     slow: mark tests as slow (deselect with '-m "not slow"')
