### 2022-07-29 9323673
if we run doctest on readme, we do not need this extra test
