### 2022-07-29 0d135f4
try to make doctests work
