### 2022-07-29 e5e306a
stdout from m3gnet made doctest upset
