+++ README.md
@@ -42,22 +42,21 @@ class with and without optional relaxation via
 # example_structures is a list of `pymatgen.core.structure.Structure` objects
 >>> from xtal2png import XtalConverter, example_structures
 >>>
->>> xc = XtalConverter(relax_on_decode=True)
->>> data = xc.xtal2png(example_structures, show=True, save=True)
->>> relaxed_decoded_structures = xc.png2xtal(data, save=False)
-assert len(relaxed_decoded_structures) == len(example_structures)
-
-```
-
-```python
->>> from xtal2png import XtalConverter, example_structures
->>>
->>> xc = XtalConverter(relax_on_decode=False)  # default behavior
+>>> xc = XtalConverter(relax_on_decode=False)
 >>> data = xc.xtal2png(example_structures, show=True, save=True)
 >>> decoded_structures = xc.png2xtal(data, save=False)
-assert len(decoded_structures) == len(example_structures)
+>>> len(decoded_structures)
+2
+
+>> xc = XtalConverter(relax_on_decode=True)
+>> data = xc.xtal2png(example_structures, show=True, save=True)
+>> relaxed_decoded_structures = xc.png2xtal(data, save=False)
+>> len(relaxed_decoded_structures)
+2
 
 ```
+
+
 
 ### Output
 
