### 2022-07-29 c71ac04
avoid repeat calculations of distance_matrix
