### 2022-07-29 sparks-baird/xtal2png#193 @sgbaird
https://github.com/sparks-baird/xtal2png/pull/197
