+++ CHANGELOG.md
@@ -1,56 +1,30 @@
 # CHANGELOG
 
-## [v0.9.1](https://github.com/sparks-baird/xtal2png/releases/tag/v0.9.1) - 2022-07-28 22:06:22
+## [v0.9.2](https://github.com/sparks-baird/xtal2png/releases/tag/v0.9.2) - 2022-07-29 16:21:51
 
-<!-- Release notes generated using configuration in .github/release.yml at v0.9.1 -->
+<!-- Release notes generated using configuration in .github/release.yml at v0.9.2 -->
 
 ## What's Changed
-* cnn-classification notebook (Created using Colaboratory) by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/175
-* add regression notebook and change classification Colab badge to markdown for compatibility with nbsphinx by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/178
-* fix: catch partial occupancy by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/182
-* fix: spacegroup fallback and installation docs  by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/186
-* feat: add `max_sites` to CLI by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/181
+* avoid repeat calculations of distance_matrix by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/199
 
 ### Other Changes
-* fixup joss badge, remove leftover \autoref command by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/172
-* split large block of README text into subsections by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/171
-* convert colab badges from html to markdown for nbsphinx compatibility by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/173
-* Revert "cnn-classification notebook (Created using Colaboratory)" by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/176
-* Created using Colaboratory by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/177
-* paper: update software paper by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/183
-* JOSS paper review - Docs by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/187
-* add Berend as co-author by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/191
-* Revert imports until conda-forge is bumped by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/189
-* response to rkurchin's editorial review by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/192
+* update authors, including adding Faris Khan by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/194
+* grant acknowledgements for Hasan and Colton by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/195
+* docs: update affiliation and acknowledgements by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/196
+* chore: test example by @kjappelbaum in https://github.com/sparks-baird/xtal2png/pull/188
+* add T.D.S. in acknowledgment statements by @sgbaird in https://github.com/sparks-baird/xtal2png/pull/197
 
 
-**Full Changelog**: https://github.com/sparks-baird/xtal2png/compare/v0.8.0...v0.9.1
-
-### Feature
-
-- general:
[…]
