### 2022-07-29 #4528 @sgbaird
Version: [`v0.9.2`](https://github.com/sparks-baird/xtal2png/tree/v0.9.2)
DOI: [10.5281/zenodo.6941615](https://doi.org/10.5281/zenodo.6941615)
