+++ reports/paper.bib
@@ -4,7 +4,7 @@
   author = {{G{\'o}mez-Bombarelli}, Rafael and Wei, Jennifer N. and Duvenaud, David and {Hern{\'a}ndez-Lobato}, Jos{\'e} Miguel and {S{\'a}nchez-Lengeling}, Benjam{\'i}n and Sheberla, Dennis and {Aguilera-Iparraguirre}, Jorge and Hirzel, Timothy D. and Adams, Ryan P. and {Aspuru-Guzik}, Al{\'a}n},
   year = {2016},
   publisher = {{arXiv}},
-  doi = {10.48550/ARXIV.1610.02415v1},
+  doi = {10.48550/ARXIV.1610.02415},
   copyright = {arXiv.org perpetual, non-exclusive license},
   keywords = {Chemical Physics (physics.chem-ph),FOS: Computer and information sciences,FOS: Physical sciences,Machine Learning (cs.LG)}
 }
@@ -60,22 +60,6 @@
   langid = {english},
   keywords = {Computer Science - Machine Learning,Condensed Matter - Materials Science,Physics - Computational Physics},
   file = {C\:\\Users\\sterg\\Zotero\\storage\\IF6WR9YR\\Goodall and Lee - 2020 - Predicting materials properties without crystal st.pdf}
-}
-
-@misc{kingmaAutoEncodingVariationalBayes2014,
-  title = {Auto-{{Encoding Variational Bayes}}},
-  author = {Kingma, Diederik P. and Welling, Max},
-  year = {2014},
-  month = may,
-  number = {arXiv:1312.6114},
-  eprint = {1312.6114},
-  eprinttype = {arxiv},
-  primaryclass = {cs, stat},
-  publisher = {{arXiv}},
-  abstract = {How can we perform efficient inference and learning in directed probabilistic models, in the presence of continuous latent variables with intractable posterior distributions, and large datasets? We introduce a stochastic variational inference and learning algorithm that scales to large datasets and, under some mild differentiability conditions, even works in the intractable case. Our contributions is two-fold. First, we show that a reparameterization of the variational lower bound yields a lower bound estimator that can be straightforwardly optimized using standard stochastic gradient methods. Second, we show that for i.i.d. datasets with continuous latent variables per datapoint, posterior inference can be made especially efficient by fitting an approximate inference model (also called a recognition model) to the intractable posterior using the proposed lower bound estimator. Theoretical advantages are reflected in experimental results.},
-  archiveprefix = {arXiv},
-  langid = {english},
-  keywords = {Computer Science - Machine Learning,Statistics - Machine Learning}
 }
 
 @misc{kingmaAutoEncodingVariationalBayes2014a,
@@ -159,22 +143,20 @@
   keywords = {data-visualization,machine-learning,materials-informatics,matplotlib,plotly,plots,uncertainty,uncertainty-calibration}
 }
 
-@misc{sahariaPaletteImagetoImageDiffusion2022,
+@inproceedings{sahariaPaletteImagetoImageDiffusion2022a,
   title = {Palette: {{Image-to-Image Diffusion Models}}},
   shorttitle = {Palette},
-  author = {Saharia, Chitwan and Chan, William and Chang, Huiwen and Lee, Chris A. and Ho, Jonathan and Salimans, Tim and Fleet, David J. and Norouzi, Mohammad},
+  booktitle = {Special {{Interest Group}} on {{Computer Graphics}} and {{Interactive Techniques Conference Proceedings}}},
[…]
