### 2022-07-29 #4528 @sgbaird
References are fixed. It looks like it didn't like the `v1` for arXiv, which is interesting since that implies the "object" associated with the DOI can still have content appended to it. Maybe that's normal though since I think Zenodo supports something similar with the latest version DOI.

Either way, working on releasing `v0.9.3`.
