### 2022-07-29 #4528 @sgbaird
Ah... woops. I thought I had changed it back to `sahariaPaletteImagetoImageDiffusion2022`, but looks like it stayed as `sahariaPaletteImagetoImageDiffusion2022a`.
